% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{addlayer}
\alias{addlayer}
\alias{addlayer,DataLayer,character-method}
\alias{layer,DataLayer-method}
\alias{layer<-,DataLayer,character-method}
\alias{layer<-,DataLayer,numeric-method}
\alias{layername,DataLayer-method}
\alias{layername<-,DataLayer,character-method}
\alias{nlayer,DataLayer-method}
\title{Layer management methods}
\usage{
addlayer(x, name)

\S4method{addlayer}{DataLayer,character}(x, name)

\S4method{layername}{DataLayer}(x)

\S4method{layername}{DataLayer,character}(x) <- value

\S4method{nlayer}{DataLayer}(x)

\S4method{layer}{DataLayer}(x)

\S4method{layer}{DataLayer,numeric}(x) <- value

\S4method{layer}{DataLayer,character}(x) <- value
}
\arguments{
\item{x}{a \code{SingleCellAssay} or derived}

\item{name}{a \code{code} character}

\item{value}{replacement}
}
\description{
A layer is a "slice" of a 3d array.  \code{SingleCellAssay} objects are stateful, in that a layer can be set, and then various routines will operate on that layer.
}
\details{
\code{addlayer(x)}: add a new layer with name \code{name}.  Returns a modified object.

\code{layername(x)}: Return name of current layer, equivalent to \code{layer(x)}.

\code{layername(x)<-value}: Rename current layer to \code{character} \code{value}

\code{nlayer(x)}: Return the number of layers.

code{layer(x)}: Return the current layer.

code{layer(x)<-value}: Change the current layer (by \code{integer} or \code{character} value)
}
\section{Methods}{

}
\examples{
data(vbetaFA)
vbetaFA <- addlayer(vbetaFA, 'expressed')
layername(vbetaFA)
ee <- exprs(vbetaFA)
layer(vbetaFA) <- 'expressed'
exprs(vbetaFA) <- (ee>0)*1
layer(vbetaFA) <- 'Et'
layername(vbetaFA) <- 'et'
layername(vbetaFA)
nlayer(vbetaFA)
}

