% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admins.R
\name{admin_match}
\alias{admin_match}
\title{Admin matching function}
\usage{
admin_match(admins, country = "SOM")
}
\arguments{
\item{admins}{Admins to match}

\item{country}{Country to match admins against. Default = "SOM"}
}
\description{
Admin matching function
}
