% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{BayesGLMlike-class}
\alias{BayesGLMlike-class}
\title{Wrapper for bayesian GLM}
\description{
Wrapper for bayesian GLM
}
\section{Slots}{

\describe{
\item{\code{prior}}{\code{numeric} optional 3d array used to specify prior for coefficients}

\item{\code{useContinuousBayes}}{\code{logical} should \code{bayesglm} be used to fit the continuous component as well?}
}}

