% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ZlmFit-bootstrap.R
\name{bootVcov1}
\alias{bootVcov1}
\title{Bootstrap a zlmfit}
\usage{
bootVcov1(zlmfit, R = 999)
}
\arguments{
\item{zlmfit}{class \code{ZlmFit}}

\item{R}{number of bootstrap replicates}
}
\value{
array of bootstrapped coefficients
}
\description{
Sample cells with replacement to find bootstrapped distribution of coefficients
}

