\name{burdenOfFiltering}
\alias{burdenOfFiltering}
\title{Burden of filtering}
\usage{
burdenOfFiltering(sc, groups, byGroup = FALSE, filt_control = NULL)
}
\arguments{
  \item{sc}{SingleCellAssay or derived class}

  \item{groups}{the groups by which to filter}

  \item{byGroup}{logical indicating whether to filter by
  group}

  \item{filt_control}{a list of control parameters.}
}
\description{
what proportions of wells are filtered due to different
criteria
}

