% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{causes_of_death}
\alias{causes_of_death}
\title{Causes of Death}
\format{
A data.frame of "Causes" and "code"
}
\usage{
causes_of_death
}
\description{
A data frame of listed causes of death in SMART files and their code
}
\keyword{datasets}
