% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{cData}
\alias{cData}
\alias{cData<-,SingleCellAssay-method}
\title{Replace cData}
\usage{
\S4method{cData}{SingleCellAssay}(sc) <- value
}
\arguments{
\item{sc}{SingleCellAssay}

\item{value}{AnnotatedDataFrame or data.frame}
}
\description{
Replace cData
}

