\docType{methods}
\name{cData}
\alias{cData}
\alias{cData,SingleCellAssay-method}
\title{cData}
\usage{
cData(sc)

\S4method{cData}{SingleCellAssay}(sc)
}
\arguments{
  \item{sc}{An object with \code{cellData}}
}
\value{
\code{data.frame}
}
\description{
Accessor for cellData \code{data.frame}
}
\details{
Returns the \code{cellData} \code{data.frame}.
}
\section{Methods}{
  \describe{ \item{\code{sc = "SingleCellAssay"}}{} }
}

\section{Methods}{
  \describe{ \item{\code{sc = "SingleCellAssay"}}{} }
}
\keyword{accessor}

