% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAssay-methods.R
\docType{methods}
\name{cData,SingleCellAssay-method}
\alias{cData,SingleCellAssay-method}
\title{Deprecated cell/feature data accessors/mutators}
\usage{
\S4method{cData}{SingleCellAssay}(sc)
}
\arguments{
\item{sc}{SingleCellAssay}

\item{value}{replacement value}
}
\value{
a \code{DataFrame} or modifies the \code{SingleCellAssay} object in place
}
\description{
These functions are now all deprecated and will be removed in a future release.
}
\section{Replacements}{


\describe{
  \item{\code{cData}}{\link{colData}}
  \item{\code{fData}}{\link{mcols}}
  \item{\code{exprs}}{\link{assay}}
  \item{\code{combine}}{\link{cbind2} or \link{rbind2}}
}
}

