\docType{methods}
\name{cellData}
\alias{cData<-}
\alias{cellData}
\alias{cellData,SingleCellAssay-method}
\title{cData}
\arguments{
  \item{sc}{An object with \code{cellData}}
}
\value{
  \code{AnnotatedDataFrame}
}
\description{
  Accessor for cellData \code{AnnotatedDataFrame}
}
\details{
  Returns the \code{cellData}.
}
\keyword{accessor}

