% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{cellData}
\alias{cellData}
\alias{cellData,SingleCellAssay-method}
\title{Accessor for cellData \code{AnnotatedDataFrame}}
\usage{
cellData(sc)
}
\arguments{
\item{sc}{An object with \code{cellData}}
}
\value{
\code{AnnotatedDataFrame}
}
\description{
Returns the \code{cellData}.
}

