% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsnau.R
\name{chr_to_num}
\alias{chr_to_num}
\title{Convert character number to numeric}
\usage{
chr_to_num(x)
}
\arguments{
\item{x}{Character vector to convert}
}
\value{
Numeric vector
}
\description{
Convert character number to numeric
}
\details{
If argument is a character, removes commas and converts
}
