% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{coef}
\alias{coef}
\title{Coefficients of zero-infated}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{LMlike}

\item{...}{passed to methods}

\item{which}{character vector, one of "C" (continuous) or "D" (discrete) specifying which component should be returned}
}
\value{
numeric vector
}
\description{
Given a fitted LMlike, return the coefficients from discrete or continuous
}

