% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAssay-methods.R
\docType{methods}
\name{colData<-,SingleCellAssay,DataFrame-method}
\alias{colData<-,SingleCellAssay,DataFrame-method}
\title{Replace \code{colData}}
\usage{
\S4method{colData}{SingleCellAssay,DataFrame}(x) <- value
}
\arguments{
\item{x}{\code{SingleCellAssay}}

\item{value}{\code{DataFrame}}
}
\value{
modified \code{SingleCellAssay}
}
\description{
Replace \code{colData} with a \code{DataFrame}.
Checks to make sure that \code{row.names(value)} match \code{colnames{x}}, in contrast to the parent method
 Checks for a wellKey column, as well.
}
