% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SingleCellAssay-methods.R
\docType{methods}
\name{combine,SCASet,missing-method}
\alias{combine,SCASet,missing-method}
\title{Combines a SCASet into a unified SingleCellAssay}
\usage{
\S4method{combine}{SCASet,missing}(x, y, ...)
}
\description{
No error checking is currently done to insure that objects in the SCASet conform with each other,
so mysterious errors may result if they do not.
}

