% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{combine,SingleCellAssay,ANY-method}
\alias{combine,SingleCellAssay,ANY-method}
\title{Combine a SingleCellAssay and a vector, data.frame or AnnotatedDataFrame}
\usage{
\S4method{combine}{SingleCellAssay,ANY}(x, y, ...)
}
\arguments{
\item{x}{SingleCellAssay}

\item{y}{vector, data.frame or AnnotatedDataFrame}

\item{...}{ignored}
}
\value{
SingleCellAssay
}
\description{
When \code{x} is a SingleCellAssay and \code{y} is a  vector, data.frame or AnnotatedDataFrame,
an attempt is made bind it to the columns of the cellData or featureData.
The behavior depends on the number of rows/length of \code{y}.
If \code{nrow(y) == nrow(x)}, then the cellData is used.
If \code{nrow(y) == ncol(x)}, then the cellData is used.
It is an error if neither mathces.
}

