% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{combine,SingleCellAssay,SingleCellAssay-method}
\alias{combine,DataLayer,Datalayer-method}
\alias{combine,SingleCellAssay,SingleCellAssay-method}
\title{Combine two SingleCellAssay or derived classes}
\usage{
\S4method{combine}{SingleCellAssay,SingleCellAssay}(x, y, ...)
}
\arguments{
\item{x}{One of the values.}

\item{y}{A second value.}

\item{...}{Any other objects of the same class as \code{x} and \code{y}.}
}
\description{
Combines two Single Cell-like objects provided they have the same number of Features and Layers.
The union of columns from featureData will be taken
The union (padded if necessary with NA) will be taken from cellData.
}
\note{
We might also wish to combine features along each cell/row but a use case for this
hasn't arrived yet.
}

