% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zlmHooks.R
\name{combined_residuals_hook}
\alias{combined_residuals_hook}
\title{Hook to get the combined residuals}
\usage{
combined_residuals_hook(x)
}
\arguments{
\item{x}{the ZLMFit}
}
\description{
Hook to get the combined residuals
}

