% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fluidigm-methods.R
\name{condmean}
\alias{condmean}
\title{Report the mean et value for each gene}
\usage{
condmean(sc)
}
\arguments{
\item{sc}{SingleCellAssay}
}
\value{
vector of means
}
\description{
NAs are always removed
}

