% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fluidigm-methods.R
\name{condSd}
\alias{condSd}
\title{Report standard deviation of et, for positive et for each gene}
\usage{
condSd(sc)
}
\arguments{
\item{sc}{SingleCellAssay}
}
\value{
vector of standard deviations
}
\description{
NAs are always removed
}

