% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zlmHooks.R
\name{continuous_residuals_hook}
\alias{continuous_residuals_hook}
\title{Hook to get the discrete residuals}
\usage{
continuous_residuals_hook(x)
}
\arguments{
\item{x}{the ZLMFit}
}
\description{
Hook to get the discrete residuals
}

