% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAssay-methods.R
\docType{methods}
\name{magic_assay_names}
\alias{magic_assay_names}
\alias{assay,SingleCellAssay,missing-method}
\alias{defaultAssay}
\title{Default \code{assay} returned}
\usage{
magic_assay_names()

\S4method{assay}{SingleCellAssay,missing}(x, i, ...)
}
\description{
Methods in this package operate on log-transformed (multiplicative scale) expression.
We attempt to check for this at construction, and then over-ride the \code{assay} method to return the "layer" containing such log-transformed data.
By default we return the assay whose names, as given by \code{assayNames(x)}, matches the first element in the vector \code{c('thresh', 'et', 'Et', 'lCount', 'logTPM', 'logCounts')}.
}
\section{Functions}{
\itemize{
\item \code{magic_assay_names}: list of names assumed to represent log-transformed data, in order of usage preference
}}

\examples{
data(vbetaFA)
assay(vbetaFA)[1:3,1:3]
assay(vbetaFA, 'thresh') = assay(vbetaFA)*0 - 9 
assay(vbetaFA)[1:3, 1:3]
}
