% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{defaultPrior}
\alias{defaultPrior}
\title{Initialize a prior to be used a prior for BayeGLMlike/BayesGLMlike2}
\usage{
defaultPrior(names)
}
\arguments{
\item{names}{character vector of coefficients.  The `(Intercept)` will be ignored.}
}
\value{
3d array, with leading dimension giving the prior 'loc'ation, 'scale' and degrees of freedom (df),
second dimension giving the component ('C'ontinuous or 'D'iscrete)
and trailing dimension giving the coefficient to which the prior applies.
The location is initialized to be 0, the scale to 2, and degrees of freedom of 1, following the default of bayesglm.
}
\description{
Initialize a prior to be used a prior for BayeGLMlike/BayesGLMlike2
}
\examples{
dp <- defaultPrior('Stim.ConditionUnstim')
\dontrun{
data(vbetaFA)
zlmVbeta <- zlm(~ Stim.Condition, vbeta.sc, method='bayesglm', coefPrior=dp)
}
}
