% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{dof}
\alias{dof}
\alias{dof,GLMlike-method}
\title{Degrees of freedom of Zero inflated model}
\usage{
dof(object)
}
\arguments{
\item{object}{LMlike or subclass}
}
\value{
vector giving the model degrees of freedom for continuous and discrete
}
\description{
Degrees of freedom of Zero inflated model
}
