% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Drop}
\alias{Drop}
\title{Drop specified dimension from an array}
\usage{
Drop(x, d)
}
\arguments{
\item{x}{array of at least d dimensions}

\item{d}{dimension(s) to drop}
}
\value{
array x
}
\description{
Like drop(x) but only dropping specified dimensions.
There is no testing that the specified dimensions are actually singletons.
}
\examples{
x = array(1:4, dim=c(1, 2, 1, 2))
dx = Drop(x, 1)
stopifnot(all(dim(dx)==c(2,1,2)))
}

