% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{estimateZlmTheta}
\alias{estimateZlmTheta}
\alias{reduceSampleInverseSE}
\alias{reduceWeightedMean}
\title{Estimate parameters for power studies from pilot data}
\usage{
estimateZlmTheta(Formula, ZlmFit, newdata, newformula,
  reduceMethod = reduceWeightedMean)

reduceWeightedMean(coef, secoef)

reduceSampleInverseSE(coef, secoef)
}
\arguments{
\item{Formula}{a \code{formula} evaluated according to a special syntax. Each term must be enclosed by \code{est} or \code{fixed}. See Details.}

\item{ZlmFit}{a fitted object that must at least contain terms that are to be `est`imated.}

\item{newdata}{an optional model frame used to construct the parameter vector.  If omitted then \code{ZlmFit} is used.  If specified, then so must \code{newformula}.}

\item{newformula}{an optional formula used to construct the parameter vector.   If omitted then \code{ZlmFit} is used.  If specified, then so must \code{newdata}.}

\item{reduceMethod}{a function, accepting M x P matrices of coefficients and standard errors, returning a px1 vector of parameters.  Considerable flexibility is possible by having reduceMethod return an unevaluated expression.  See \link{reduceSampleInverseSE} for an example.}

\item{coef}{coefficient matrix}

\item{secoef}{standard error matrix}
}
\value{
a list of class ZlmTheta, containing components `C`, `D` and `sigmaC`.
}
\description{
This function is intended to facilitate the use of pilot/related studies to fix some parameters in a power analysis.  It returns a list of single cell expression parameters from a combination of pilot data, in the form of a fitted zlm object \code{ZlmFit} and fixed parameters using a formula-like interface.
}
\details{
Each term in the \code{Formula} must be enclosed by a functional keyword. The functional keyword \code{est(term)}, requests estimation of continuous/discrete expression parameters specified in \code{term} from pilot data.  Hence \code{term} must be present in the formula used to fit \code{ZlmFit}.  If \emph{term} names a factor that was expanded in terms of multiple contrasts, then each contrast will be estimated.

 The functional keyword \code{fixed(contrast, C=continuous_parameter, D=discrete_parameter)}, fixes the continuous/discrete parameters for the particular \emph{contrast} to specified values.  So if \code{contrast} specifies a term from a factor, it should be named according to the contrast applied.  See the example.

Either \code{term} or \code{contrast} can be set to the special values `1`, which is the intercept term, and `error`, which is the error term for the continuous regression.

Terms that were present in the ZlmFit formula (or newformula, if specified) but were omitted in the \code{Formula} are set to 0, except the continuous residual standard error is set to 1.

Since `ZlmFit` may contains many genes, the parameter values are reduced by applying \code{reduceMethod} to the coefficient vector and its standard errors.  The default method returns a weighted average, inversely weighted by the standard errors.
}
\section{Functions}{
\itemize{
\item \code{reduceWeightedMean}: reduce a matrix of coefficients and standard errors to the weighted columns means

\item \code{reduceSampleInverseSE}: sample from a matrix of coefficients, inversely weighted by the standard errors.  Returns a list of unevaluated expressions, one for each column.
}}
\examples{
data(vbetaFA)
ZlmFit <- zlm.SingleCellAssay(~Population+Subject.ID+ncells, vbetaFA, method='bayesglm')
estimateZlmTheta(~est(1) + est(Population) + fixed(Subject.IDSub02, C=0, D=.5) + est(error), ZlmFit)
}

