% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expavg}
\alias{expavg}
\title{Exponential average}
\usage{
expavg(x)
}
\arguments{
\item{x}{\code{numeric}}
}
\value{
\code{numeric}
}
\description{
Puts log transformed values onto natural scale and takes mean of vector.
Calculates mean(2^x - 1)
}

