\docType{methods}
\name{exprs}
\alias{exprs}
\alias{exprs,SingleCellAssay-method}
\alias{exprs<-,SingleCellAssay,ANY-method}
\title{exprs}
\arguments{
  \item{object}{SingleCellAssay}
}
\value{
  numeric matrix
}
\description{
  Get or set a matrix of measurement values in a
  \code{SingleCellAssay}
}
\details{
  Return or set a matrix of the measurement: cells by
  primerids
}

