% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataLayer.R
\docType{methods}
\name{exprsLayer,DataLayer,numeric-method}
\alias{exprsLayer,DataLayer,numeric-method}
\alias{exprsLayer-DataLayer-character}
\title{Get or set a matrix of measurement values in a \code{SingleCellAssay}}
\usage{
\S4method{exprsLayer}{DataLayer,numeric}(object, layernm)
}
\arguments{
\item{object}{DataLayer}

\item{layernm}{optional \code{character} or \code{numeric} giving the expression layer to be returned}
}
\value{
a \code{matrix} of measurement values with wells on the rows and features on the columns of the default layer
}
\description{
Return or set a matrix of the measurement: cells by primerids
}

