% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{exprsLayer<-}
\alias{exprsLayer<-}
\title{Set an Expression matrix by layername}
\usage{
exprsLayer(object, layernm, ...) <- value
}
\arguments{
\item{object}{DataLayer}

\item{ANY}{layernm}
}
\value{
matrix
}
\description{
Set an Expression matrix by layername
}

