% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{exprsLayer}
\alias{exprsLayer}
\title{Get Expression matrix by layername}
\usage{
exprsLayer(object, layernm, ...)
}
\arguments{
\item{object}{DataLayer}

\item{ANY}{layernm}
}
\value{
matrix
}
\description{
Get Expression matrix by layername
}

