% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{f_calc_days}
\alias{f_calc_days}
\title{Function to clean metadata smart surveys}
\usage{
f_calc_days(f_surveys_cov, f_df)
}
\arguments{
\item{f_surveys_cov}{Surveys Coverage}

\item{f_df}{dataframe}
}
\description{
Function to clean metadata smart surveys
}
\details{
Function to clean metadata smart surveys
}
