% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{f_calculate_weighted_predictors}
\alias{f_calculate_weighted_predictors}
\title{Function to join smart surveys and predictors}
\usage{
f_calculate_weighted_predictors(
  predictors_data,
  ts,
  surveys_cov,
  hh_obs_complete,
  col_admin2
)
}
\arguments{
\item{predictors_data}{Predictor Datasets}

\item{ts}{dataframe contraining time units and admin2}

\item{surveys_cov}{Coverage of the SMART Surveys}

\item{hh_obs_complete}{Data Frame of the COMPLETE SMART SURVEYS}

\item{col_admin2}{name of the admin2 column (most of the time 'district')}
}
\description{
Function to join smart surveys and predictors
}
\details{
Function to join smart surveys and predictors by weights
}
