% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{f_categorise}
\alias{f_categorise}
\title{Transform Variables into Categorical Variables}
\usage{
f_categorise(data, breaks_and_labels)

f_categorise(data, breaks_and_labels)
}
\arguments{
\item{data}{DataFrame}

\item{breaks_and_labels}{predictors and there breaks}
}
\value{
\link{data.frame} DataFrame with the addition of categorical variables

\link{data.frame} DataFrame completed
}
\description{
Transform Variables into Categorical Variables

Create categories
}
\details{
Modify the variable into transformed variables

Modify the dataset in order to categorize some variables
}
