% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{f_clean_metadata}
\alias{f_clean_metadata}
\title{Function to clean metadata smart surveys}
\usage{
f_clean_metadata(metadata, ts)
}
\arguments{
\item{metadata}{SMART Surveys Metadata}

\item{ts}{dataframe contraining time units and admin2}
}
\description{
Function to clean metadata smart surveys
}
\details{
Function to clean metadata smart surveys
}
