% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{f_complete_mice}
\alias{f_complete_mice}
\title{Create complete dataset by using mice}
\usage{
f_complete_mice(data, variable, col_admin2)
}
\arguments{
\item{data}{DataFrame}

\item{variable}{Variable to impute}

\item{col_admin2}{colnames of the admin2 level}
}
\value{
\link{data.frame} DataFrame completed
}
\description{
Create complete dataset by using mice
}
\details{
Modify the dataset in order to complete the variables
}
