% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{f_create_lags}
\alias{f_create_lags}
\title{Create lags}
\usage{
f_create_lags(data, vars, col_admin2, col_admin1, lags = 1:6)
}
\arguments{
\item{data}{DataFrame}

\item{vars}{predictors to be lags}

\item{col_admin2}{colnames of the admin2 level}

\item{col_admin1}{colnames of the admin1 level}

\item{lags}{lagged time}
}
\value{
\link{data.frame} DataFrame completed
}
\description{
Create lags
}
\details{
Modify the dataset in order to lags some variable
}
