% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_units.R
\name{f_gen_ts}
\alias{f_gen_ts}
\title{Generate time units cross district dataframe}
\usage{
f_gen_ts(
  admin2_f,
  name_col,
  y_start = 2014,
  y_end = 2023,
  m_start = 1,
  m_end = 12,
  burn_in_period = 1,
  burn_out_period = 0
)
}
\arguments{
\item{admin2_f}{district dataset containing all the different admin2 level of the country}

\item{y_start}{Year study started. Defaults 2014.}

\item{y_end}{Year study ended. Defaults 2023}

\item{m_start}{Month study started. Defaults 1.}

\item{m_end}{Month study ended. Defaults 12}

\item{burn_in_period}{by how many years to extend period backwards. Defaults 1.}

\item{burn_out_period}{by how many years to extend period forward.Defaults 0.}
}
\value{
\link{data.frame} cross join time units and admin2 levels
}
\description{
Generate time units cross district dataframe
}
\details{
Generate time unit cross district dataframe
}
