% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{f_norm_preds}
\alias{f_norm_preds}
\title{Create scale and normalize predictors}
\usage{
f_norm_preds(data, preds_scn)
}
\arguments{
\item{data}{DataFrame}

\item{preds_scn}{predictors to be normalized}
}
\value{
\link{data.frame} DataFrame completed
}
\description{
Create scale and normalize predictors
}
\details{
Modify the dataset in order to norm some variable
}
