% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{f_plot_missing_data}
\alias{f_plot_missing_data}
\title{Function to plot the proportion of missing predictors}
\usage{
f_plot_missing_data(data)
}
\arguments{
\item{data}{DataFrame to plot}
}
\description{
Function to plot the proportion of missing predictors
}
\details{
Function to plot the proportion of missing predictors
}
