% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{f_rate_variables}
\alias{f_rate_variables}
\title{Create rate dataset using the population data}
\usage{
f_rate_variables(data, pred_rates, col_admin2, col_pop)
}
\arguments{
\item{data}{DataFrame}

\item{pred_rates}{predictors to be rated}

\item{col_admin2}{colnames of the admin2 level}

\item{col_pop}{colnames of the population}
}
\value{
\link{data.frame} DataFrame completed
}
\description{
Create rate dataset using the population data
}
\details{
Modify the dataset in order to rate some variables
}
