% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqAssay-methods.R
\name{filterLowExpressedGenes}
\alias{filterLowExpressedGenes}
\title{Filter low-expressing genes}
\usage{
filterLowExpressedGenes(assay, threshold = 0.1)
}
\arguments{
\item{assay}{a \code{SingleCellAssay} object}

\item{threshold}{a \code{numeric} between 0, and 1, specifying the threshold frequency below which genes will be filtered out}
}
\value{
\code{SingleCellAssay}
}
\description{
Filter out genes that have less than some percent threshold expression across all libraries
}
\examples{
data(vbetaFA)
filterLowExpressedGenes(vbetaFA)
}
