% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmWrapper-glmer.R
\docType{methods}
\name{fit,LMERlike,missing-method}
\alias{fit,LMERlike,missing-method}
\title{##' @include AllClasses.R}
\usage{
\S4method{fit}{LMERlike,missing}(object, response, silent = TRUE, ...)
}
\description{
##' @include AllClasses.R
}

