% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{fit}
\alias{fit}
\alias{fit,GLMlike,missing-method}
\title{fit a zero-inflated regression}
\usage{
fit(object, response, ...)
}
\arguments{
\item{object}{inheriting from \code{LMlike}}

\item{response}{a vector, same length as the design, or if missing then use the current response}

\item{...}{currently ignored}
}
\value{
LMlike or subclass
}
\description{
Given a design and formula, fit the zero inflated regression, storing the fits in slots
\code{fitC} and \code{fitD}
}

