% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsnau.R
\name{fsnau_dashboard}
\alias{fsnau_dashboard}
\title{Download FSNAU Dashboard}
\usage{
fsnau_dashboard(years = 2013:2022)
}
\arguments{
\item{years}{Years to fetch. Default = 2013:2022}
}
\value{
\link{data.frame} of all requested FSNAU year data
}
\description{
Download FSNAU Dashboard
}
\details{
Downloads the FSNAU Dashboard available at
\href{https://dashboard.fsnau.org/dashboard}{FSNAU Dashboard}
for the years requested.
}
