\docType{methods}
\name{getcellKey-method}
\alias{getcellKey}
\alias{getcellKey,SingleCellAssay-method}
\alias{getcellKey-method}
\title{getcellKey}
\arguments{
  \item{sc}{An object with a \code{cellKey}}
}
\value{
  \code{integer} giving the unique id generated
}
\description{
  Accessor for cellKey
}
\details{
  This returns the cellKey, which is a unique identifier
  generated by \code{idvars} in the \code{mapping}
}
\keyword{accessor}

