\name{getConcordance}
\alias{getConcordance}
\title{getConcordance}
\usage{
  getConcordance(singleCellRef, singleCellcomp,
    groups = NULL, fun.natural = expavg,
    fun.cycle = logmean)
}
\arguments{
  \item{singleCellRef}{}

  \item{singleCellcomp}{}

  \item{groups}{}

  \item{fun.natural}{}

  \item{fun.cycle}{}
}
\value{
  concordance between two assays
}
\description{
  Get the concordance between two
}
\details{
  Return the concordance between two assays (i.e. single
  cell and hundred cell)
}
\author{
  Andrew McDavid
}

