% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{getLambda}
\alias{getLambda}
\title{Simulate tests on a single cell gene expression data set using hurdle model}
\usage{
getLambda(nsim, thetafun, Xfun, Formula, contrasts, zlmargs = NULL)
}
\arguments{
\item{nsim}{number of simulations to run}

\item{thetafun}{function returning a realization of the parameters (return a constant if parameters are fixed)}

\item{Xfun}{function returning a realization of the design matrix.  Must explicitly include an intercept.}

\item{Formula}{formula used to model the design}

\item{contrasts}{argument passed to `lrTest`}

\item{zlmargs}{arguments passed to `zlm.SingleCellAssay`}
}
\value{
matrix containing likelihood ratio stats, p values and dof
}
\description{
Simulate tests on a single cell gene expression data set using hurdle model
}
\examples{
getLambda(100, thetafun, Xfun=simulateX(.7, 3, 96), ~hiv+(1|ind), 'hiv', zlmargs=list(method='glmer'))
}

