\docType{methods}
\name{getMapNames}
\alias{getMapNames}
\alias{getMapNames,Mapping-method}
\alias{getMapNames,SCA-method}
\title{getMapNames}
\arguments{
  \item{object}{An object with a \code{Mapping} or
  \code{mapNames} slot}
}
\value{
  character vector of map names
}
\description{
  Accessor for mapNames

  ##'
}
\details{
  This returns the mapNames, which are fields recognized in
  the names of mapping
}
\keyword{accessor}

