% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fluidigm-methods.R
\name{getwss}
\alias{getrc}
\alias{getss}
\alias{getwss}
\title{Concordance correlation coefficient lin 1989}
\usage{
getwss(concord, nexp)

getss(concord)

getrc(concord)
}
\arguments{
\item{concord}{data.frame returned by getConcordance}

\item{nexp}{number of expressed cells per row in \code{concord}}
}
\value{
numeric
}
\description{
Return's Lin's intraclass correlation coefficient for concordance
}
\section{Functions}{
\itemize{
\item \code{getwss}: the sum of squares, weighted by nexp

\item \code{getss}: the sum of squares
}}

