% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Fluidigm-methods.R
\name{getss}
\alias{getss}
\title{getss}
\usage{
getss(concord)
}
\arguments{
\item{concord}{output from \code{getConcordance}}
}
\value{
sum of squares
}
\description{
getss
}
\note{
Ditto.. compute on the natural scale
}

