% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fluidigm-methods.R
\name{getwss}
\alias{getwss}
\title{getwss}
\usage{
getwss(concord, nexp)
}
\arguments{
\item{concord}{output from \code{getConcordance}}

\item{nexp}{number of expressed cells per row in \code{concord}}
}
\value{
weighted sum of squares
}
\description{
getwss
}

