% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GSEATests-class}
\alias{GSEATests-class}
\title{An S4 class for Gene Set Enrichment output}
\description{
This holds output from a call to gseaAfterBoot.
It primarily provides a summary method.
}
\section{Slots}{

\describe{
\item{\code{tests}}{array: gene sets X {discrete,continuous} X {stat, variance, degrees of freedom, avg correlation} X {test, null}}

\item{\code{bootR}}{number of bootstrap replicates}
}}

\seealso{
gseaAfterBoot

calcZ

summary,GSEATests-method
}
