% Generated by roxygen2 (4.0.1): do not edit by hand
\name{guessContrast}
\alias{guessContrast}
\title{try to guess form of LHT}
\usage{
guessContrast(hypo.terms, model)
}
\arguments{
\item{hypo.terms}{}

\item{model}{}
}
\value{
character vector of terms to test
}
\description{
try to guess form of LHT
}

