% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hushWarning}
\alias{hushWarning}
\title{Selectively muffle warnings based on output}
\usage{
hushWarning(expr, regexp)
}
\arguments{
\item{expr}{an expression}

\item{regexp}{a regexp to be matched (with str_detect)}
}
\value{
the result of expr
}
\description{
Selectively muffle warnings based on output
}

