% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{impute}
\alias{impute}
\title{impute missing continuous expression for plotting}
\usage{
impute(object, groupby)
}
\arguments{
\item{object}{Output of predict}

\item{groupby}{Variables (column names in predict) to group by for imputation (facets of the plot)}
}
\value{
data.table
}
\description{
If there are no positive observations for a contrast, it is generally not estimible.
However, for the purposes of testing we can replace it with the least favorable value with respect to the contrasts that are defined.
}
\examples{
##See stat_ell
example(stat_ell)
}
