% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zlmHooks.R
\name{influence.bayesglm}
\alias{influence.bayesglm}
\title{Influence bayesglm object}
\usage{
\method{influence}{bayesglm}(model, do.coef = TRUE, ...)
}
\arguments{
\item{model}{\code{bayesglm}}

\item{do.coef}{see \link{influence.glm}}

\item{...}{ignored}
}
\value{
see \link{influence.glm}
}
\description{
The influence function
}
