% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesglm.R
\name{invlogit}
\alias{invlogit}
\title{Inverse of logistic transformation}
\usage{
invlogit(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
Inverse of logistic transformation
}
\examples{
x <- 1:5
invlogit(log(x/(1-x)))
}
