\docType{methods}
\name{isEmtpy}
\alias{isEmpty}
\alias{isEmpty,Mapping-method}
\alias{isEmtpy}
\title{iEmpty}
\arguments{
  \item{object}{The \code{Mapping}}
}
\value{
  A \code{logical}. \code{TRUE} if empty. \code{FALSE}
  otherwise.
}
\description{
  Is this Mapping Empty
}
\details{
  Tests if a Mapping is empty
}

