% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-methods.R
\name{jailBreak}
\alias{jailBreak}
\title{Return a dependency-free version of a SingleCellAssay Object}
\usage{
jailBreak(sca, file)
}
\arguments{
\item{sca}{\code{SingleCellAssay} object}

\item{file}{optional \code{character} naming a file to which a .RDS will be written}
}
\value{
a list (invisibly if \code{file} is set)
}
\description{
Breaks \code{sca} into a list of components `exprs` (expression matrix, cell X genes),
`cdata` (\code{data.frame} of cell annotations)
`fdata` (\code{data.frame} of feature annotations).
Optionally writes this list to an RDS file.
}

