% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ken_admin2}
\alias{ken_admin2}
\title{Kenya Admin 2 Names with counties and code}
\format{
A data.frame of counties, subcounties and code
}
\usage{
ken_admin2
}
\description{
A data frame of admin with counties, sub counties and code
}
\keyword{datasets}
