% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ken_admins}
\alias{ken_admins}
\title{Kenya Admin Names}
\format{
A data.frame of "wrong" (all names) and "right" (names to use)
}
\usage{
ken_admins
}
\description{
A data frame of admin names to match against
}
\keyword{datasets}
