\docType{methods}
\name{lapply}
\alias{lapply}
\alias{lapply,SCASet-method}
\title{lapply}
\usage{
\S4method{lapply}{SCASet}(X, FUN, ...)
}
\arguments{
  \item{X}{ A list-like or vector-like R object.  }

  \item{FUN}{ See \code{?base::\link[base]{lapply}} for a
  description of these arguments.  }
}
\description{
lapply
}

