% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{layer}
\alias{layer}
\title{Which layer is active?}
\usage{
layer(x)
}
\arguments{
\item{x}{DataLayer}
}
\value{
numeric, active layer
}
\description{
Which layer is active?
}
\seealso{
addlayer
}

