% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{layername<-}
\alias{layername<-}
\title{Set name of active layer}
\usage{
layername(x) <- value
}
\arguments{
\item{x}{DataLayer}

\item{value}{character}
}
\value{
DataLayer
}
\description{
Set name of active layer
}
\seealso{
addlayer
}

