% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/lmWrapper-glmer.R
\docType{class}
\name{LMERlike-class}
\alias{LMERlike-class}
\alias{coef,LMERlike-method}
\alias{vcov,LMERlike-method}
\title{Wrapper for lmer/glmer}
\usage{
\S4method{vcov}{LMERlike}(object, which, ...)

\S4method{coef}{LMERlike}(object, which, singular = TRUE, ...)
}
\arguments{
\item{object}{\code{LMERlike}}

\item{which}{\code{character}, one of 'C', 'D'.}

\item{...}{ignored}

\item{singular}{\code{logical}. Should NA coefficients be returned?}
}
\description{
A horrendous hack is employed in order to do arbitrary likelihood ratio tests: the model matrix is built, the names possibly mangled, then fed in as a symbolic formula to glmer/lmer.
This is necessary because there is no (easy) way to specify an arbitrary fixed-effect model matrix in glmer.
}
\section{Methods (by generic)}{
\itemize{
\item \code{vcov}: return the variance/covariance of component \code{which}

\item \code{coef}: return the coefficients.  The horrendous hack is attempted to be undone.
}}
\section{Slots}{

\describe{
\item{\code{pseudoMM}}{part of this horrendous hack.}
}}

