\name{logmean}
\alias{logmean}
\title{Log mean}
\usage{
  logmean(x)
}
\arguments{
  \item{x}{\code{numeric}}
}
\value{
  \code{numeric}
}
\description{
  Takes mean of natural scaled values and then logrithm
  Approximately the inverse operation of
  \code{\link{expavg}} Calculates log2(mean(x) + 1)
}

