% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZlmFit.R
\docType{methods}
\name{lrTest,ZlmFit,CoefficientHypothesis-method}
\alias{lrTest,ZlmFit,CoefficientHypothesis-method}
\title{@describeIn ZlmFit Returns an array with likelihood-ratio tests on contrasts defined using \code{CoefficientHypothesis()}.}
\usage{
\S4method{lrTest}{ZlmFit,CoefficientHypothesis}(object, hypothesis)
}
\description{
@describeIn ZlmFit Returns an array with likelihood-ratio tests on contrasts defined using \code{CoefficientHypothesis()}.
}

