% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lrTest}
\alias{lrTest}
\title{Run a likelihood-ratio test}
\usage{
lrTest(object, drop.terms)
}
\arguments{
\item{object}{LMlike or subclass}

\item{drop.terms}{character vector of \code{formula} terms}
}
\value{
array giving test statistics
}
\description{
Compares the change in likelihood between the current \code{formula} and one dropping terms in \code{drop.terms}.
Only complete terms can be tested at this time
}
\seealso{
fit

waldTest
}

