% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{maits}
\alias{maits}
\title{MAITs data set, RNASeq}
\format{a \code{list} containing an expression matrix (\code{expressionmat}), cell \code{cdat} and feature \code{fdat}.}
\description{
MAITs data set, RNASeq
}
\seealso{
\code{\link{FromMatrix}}
}
