% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST-package.R
\name{MAST-defunct}
\alias{MAST-defunct}
\alias{filter}
\title{Defunct functions in package `MAST`}
\description{
These functions are defunct or have been renamed.
}
\section{Functions (and replacements, if available)}{

\describe{
  \item{filter}{mast_filter}
}
}

