% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAssay-methods.R
\docType{package}
\name{MAST-package}
\alias{MAST-package}
\title{"Tools for Single Cell Assay Analysis"}
\description{
Methods for analysing single cell assay data
}
\details{
This packages provides data structures and functions for statistical analysis of single-cell assay data such as Fluidigm single cell gene expression assays.
}
\keyword{package}

