% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mast.R
\docType{package}
\name{mast}
\alias{mast}
\alias{mast-package}
\title{\pkg{mast} Munging and samrat tools}
\description{
Package for data munging (cleaning, collating etc) related
to the samrat package.
}
\author{
\strong{Maintainer}: OJ Watson \email{oj.watson@hotmail.co.uk} (\href{https://orcid.org/0000-0003-2374-0741}{ORCID})

}
\keyword{internal}
