% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{mast_excel_lists}
\alias{mast_excel_lists}
\title{mast_excel_lists}
\usage{
mast_excel_lists(path, sheets = NULL, pattern = NULL, ...)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{sheets}{Name of sheets in xlsx file. Default = \code{NULL}. Takes
priority over regex}

\item{pattern}{Regex for sheets to be used. Default = \code{NULL}}

\item{...}{Other arguments passed to \link[readxl:read_excel]{readxl::read_xlsx}}
}
\description{
Reads sheets from excel file and converts
into a list, with named sheet per list.
}
\details{
\code{mast_excel_lists}
}
