% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-methods.R
\name{meld_list_left}
\alias{meld_list_left}
\title{Combine lists, preferentially taking elements from x if there are duplicate names}
\usage{
meld_list_left(x, y)
}
\arguments{
\item{x}{list}

\item{y}{list}
}
\description{
Combine lists, preferentially taking elements from x if there are duplicate names
}
\examples{
MAST:::meld_list_left(list(A=1, B=2), list(A = 0))
}
