% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NanostringRCCReader.R
\name{mergeWithKeyFile}
\alias{mergeWithKeyFile}
\title{Merge NanoString lanes with a key file}
\usage{
mergeWithKeyFile(rcc, file)
}
\arguments{
\item{rcc}{is a list returned from readNanoStringLanes}

\item{file}{is a character vector name of the key file}
}
\value{
mapped rcc list
}
\description{
Reads in a key file and maps to nanostring lanes
}

