% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{model.matrix<-}
\alias{model.matrix<-}
\title{Replace model matrix}
\usage{
model.matrix(object) <- value
}
\arguments{
\item{object}{LMlike or subclass}

\item{value}{matrix}
}
\value{
modify object
}
\description{
Replace model matrix
}
