% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{model.matrix}
\alias{model.matrix}
\title{Model matrix accessor}
\usage{
model.matrix(object, ...)
}
\arguments{
\item{object}{LMlike or subclass}
}
\value{
model.matrix if present
}
\description{
Model matrix accessor
}

