% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{nlayer}
\alias{nlayer}
\title{How many layers does a DataLayer have?}
\usage{
nlayer(x)
}
\arguments{
\item{x}{DataLayer}
}
\value{
numeric, number of layers
}
\description{
How many layers does a DataLayer have?
}
\seealso{
addlayer
}

