% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqAssay-methods.R
\name{outlierDetection}
\alias{outlierDetection}
\title{Filter an RNASeqAssay}
\usage{
outlierDetection(assay, plot = TRUE)
}
\arguments{
\item{assay}{a \code{RNASeqAssay}}

\item{plot}{a \code{logical} indicating whether plots should be generated}
}
\value{
a filtered RNASeqAssay
plot the outliers
}
\description{
Use the percentage of library per gene CDF type plots
to pick out outliers using the \code{1.5*IQR}
threshold
}

