% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZlmFit-bootstrap.R
\name{pbootVcov1}
\alias{pbootVcov1}
\title{Bootstrap a zlmfit}
\usage{
pbootVcov1(cl, zlmfit, R = 99)
}
\arguments{
\item{cl}{a \code{cluster} object created by \code{makeCluster}}

\item{zlmfit}{class \code{ZlmFit}}

\item{R}{number of bootstrap replicates}
}
\value{
array of bootstrapped coefficients
}
\description{
Sample cells with replacement to find bootstrapped distribution of coefficients
}
