% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdSCRNA.R
\name{plot.thresholdSCRNACountMatrix}
\alias{plot.thresholdSCRNACountMatrix}
\title{Plot cutpoints and densities for thresholding}
\usage{
\method{plot}{thresholdSCRNACountMatrix}(x, ask = FALSE, wait.time = 0,
  type = "bin", indices = NULL, ...)
}
\arguments{
\item{x}{output of \code{thresholdSCRNACountMatrix}}

\item{ask}{if TRUE then will prompt before displaying each plot}

\item{wait.time}{pause (in seconds) between each plot}

\item{type}{one or more of the following: 'bin' (plot the genes by the binning used for thresholding), or 'gene' (plot thresholding by gene -- see next argument)}

\item{indices}{if \code{type} is equal to 'gene', and is a integer of length 1, then a random sample of \code{indices} genes is taken.  If it is NULL, then 10 genes are sampled.  If it is a integer vector of length > 1, then it is interpreted as giving a list of indices of genes to be displayed.}

\item{...}{further arguments passed to \code{plot}}
}
\value{
displays plots
}
\description{
Plot cutpoints and densities for thresholding
}
\examples{
## See thresholdSCRNACountMatrix
example(thresholdSCRNACountMatrix)
}

