% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.R
\name{plotlrt}
\alias{plotlrt}
\title{Plot a likelihood ratio test object}
\usage{
plotlrt(lr, adjust = "fdr", thres = 0.1, trunc = 1e-06,
  groups = NULL)
}
\arguments{
\item{lr}{output from lrtest, with returnall=FALSE}

\item{adjust}{\code{character}, passed along to \code{p.adjust}, see below}

\item{thres}{\code{numeric} genes with adjusted pvalues above this value are not depicted}

\item{trunc}{\code{numeric} p values below this value are truncated at this value}

\item{groups}{\code{character} grouping value.  If provided, must match groups argument passed to lrtest.  Plots done separately for each group.}
}
\value{
Constructs a dotplot
}
\description{
Constructs a forest-like plot of signed log10 p-values, possibly adjusted for multiple comparisons
\code{adjust} can be one of  "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".
}
\author{
andrew
}
