% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ZlmFit}
\alias{predict.ZlmFit}
\title{Return predictions from a ZlmFit object.}
\usage{
\method{predict}{ZlmFit}(object, newdata = NULL, modelmatrix = NULL)
}
\arguments{
\item{object}{A \code{ZlmFit}}

\item{newdata}{The data to predict from. Currently ignored, will use the data in the object.}

\item{modelmatrix}{The model matrix specifying the linear combination of coefficients.}

\item{sepinto}{Separate the sample column into groups with these named columns}

\item{groupby}{variables to facet by for the plot, will impute missing based on facets.}
}
\value{
Predictions and standard errors.
}
\description{
Return predictions from a ZlmFit object.
}

