% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{predicted_sig}
\alias{predicted_sig}
\title{Predicted signatures}
\format{A data frame of predicted gene expresion signatures for stimulated and unstimulated cells.}
\description{
Predicted signatures
}
