% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fluidigm-methods.R
\name{primerAverage}
\alias{primerAverage}
\title{Average within duplicated genes/primers}
\usage{
primerAverage(fd, geneGroups, fun.natural = expavg, fun.cycle = logshift)
}
\arguments{
\item{fd}{\code{SingleCellAssay} or subclass}

\item{geneGroups}{\code{character} naming a column in the \code{featureData} that keys the duplicates}

\item{fun.natural}{transformation to be used to collapse the duplicate expression values}

\item{fun.cycle}{transformation to be used after collapsing}
}
\value{
collapsed version of \code{fd}.
}
\description{
.
}

