% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_mast_file}
\alias{read_mast_file}
\title{read_mast_file}
\usage{
read_mast_file(name, sheet = NULL)
}
\arguments{
\item{name}{Name of a file within the inst/extdata folder.}

\item{sheet}{Name of sheet in xlsx file. Default = \code{NULL}}
}
\description{
Load a file from within the inst/extdata folder of the
\link{mast} package. File extension must be one of .csv, .rds, or .xlsx
}
\details{
\code{read_mast_file} read mast package file
}
\keyword{internal}
