% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{.readSection}
\alias{.readSection}
\title{Read a section of a Nanostring RCC file}
\usage{
.readSection(x, file, type, recast = TRUE, ...)
}
\arguments{
\item{x}{character vector of lines from RCC file, newlines stripped}

\item{file}{name}

\item{type}{the section of the file to be read in}

\item{recast}{should the section be transposed?}

\item{...}{argument spassed to read.csv}
}
\value{
data.frame of section
}
\description{
The section is surrounded by <type> </type>
}

