% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{repeatedMeasuresAsymptoticPower}
\alias{repeatedMeasuresAsymptoticPower}
\title{Analytic/Asymptotic power calculations for repeated measures on single cells}
\usage{
repeatedMeasuresAsymptoticPower(design = "crosssectional", sig.level = 0.05,
  power = 0.8, nWithin = 50, sigWithin = sqrt(1.4), nBetween = 16,
  sigBetween = sqrt(0.1), logfc = NULL, ...)
}
\arguments{
\item{design}{one of "crosssectional" or "longitudinal"}

\item{sig.level}{level of test}

\item{power}{power of test}

\item{nWithin}{number of cells sampled per subject (per subject per treatment for longitudinal designs)}

\item{sigWithin}{average s.d. of expression level within subject (within treatment, for longitudinal designs)}

\item{nBetween}{number of subjects (per treatment, for cross-sectional designs)}

\item{sigBetween}{s.d. of treatment effect, across subjects}

\item{logfc}{overall average log-fold change between treatment A and B}

\item{...}{arguments passed to uniroot}
}
\value{
vector of power parameters
}
\description{
If groups A and B recieve treatments A and B, and there are `nBetween` subjects in each group, and each subject has `nWithin` cells sampled,
then the design is "cross-sectional"
If each individual receives treatments A and B, and there are `nBetween` subjects and `nWithin cells sampled *in each treatment, per individual, then
the design is "longitudinal"
In either case, the power of detecting a difference of size `logfc` at significance `sig.level` is provided
given within-subject standard deviation `sigWithin` (this is the average s.d. of cellular expression within in subject) and between-subject standard deviation `sigBetween` (this is the s.d. of the \emph{mean} level of expression for each subject.
}
\details{
At most one of the power parameters can omitted, by setting it to \code{NULL}--
then it will be solved in terms of the other provided parameters.

This analysis does not account for zero-inflation/bi-modality but could be useful as an lower bound on power when  \code{sigWithin} and \code{sigBetween} are set large enough (ie, set at zero-inflated levels).
}
\references{
Diggle et al 2002, chapter 2
}

