% Generated by roxygen2 (4.0.1): do not edit by hand
\name{outlierDetection}
\alias{outlierDetection}
\title{Filter an RNASeqAssay}
\usage{
outlierDetection(assay, plot = TRUE)
}
\arguments{
\item{assay}{a \code{RNASeqAssay}}

\item{plot}{a \code{logical} indicating whether plots should be generated}
}
\value{
a filtered RNASeqAssay
}
\description{
Use the % of library per gene CDF type plots
to pick out outliers using the 1.5*IQR
threshold
############################################
}

