% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zlmHooks.R
\name{rstandard.bayesglm}
\alias{rstandard.bayesglm}
\title{rstandard for bayesglm objects.}
\usage{
\method{rstandard}{bayesglm}(model, infl = influence(model, do.coef = FALSE),
  type = c("deviance", "pearson"), ...)
}
\arguments{
\item{model}{\code{bayesglm}}

\item{infl}{see \link{rstandard}}

\item{type}{see \link{rstandard}}

\item{...}{ignored}
}
\value{
\code{numeric} residuals
}
\description{
rstandard bayesglm object S3 method
}
