\docType{methods}
\name{sampleNames}
\alias{sampleNames}
\title{get the sample names in an SCASet}
\usage{
\S4method{sampleNames}{SCASet}(object)
}
\arguments{
  \item{object}{The \code{SCASet} object}
}
\value{
a \code{character} vector of the names of the samples in
the \code{SCASet}
}
\description{
get the sample names in an SCASet
}
\keyword{transform}

