% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_figs}
\alias{save_figs}
\title{Save Figures Nicely}
\usage{
save_figs(
  name,
  fig,
  width = 6,
  height = 6,
  root = file.path(here::here(), "analysis/plots")
)
}
\arguments{
\item{name}{Name of saved file (no file ending required). E.g. "my_plot"}

\item{fig}{ggplot or similar object}

\item{width}{Width of figure in inches. Default = 6}

\item{height}{Height of figure in inches. Default = 6}

\item{root}{Directory where figures are saved. Default is :
\code{file.path(here::here(), "analysis/plots")}}
}
\description{
Saves ggplot and similar figures to both a pdf and
to a png with the same size. PDF can then be easily edited after
if needed for quick or complicated figure annotation.
}
\details{
\code{save_figs} saves ggplot and similar figures to file
}
\keyword{internal}
