\docType{class}
\name{SCASet-class}
\alias{SCASet-class}
\title{SCASet is a set of SingleCellAssay objects or objects of its subclasses (i.e. FluidigmAssay)
The constructor \code{SCASet} should be used to make objects of this class.
}
\section{Slots}{
\describe{
\item{set}{A \code{list} of \code{SingleCellAssays} or its subclasses}
}}
\description{
SCASet is a set of SingleCellAssay objects or objects of
its subclasses (i.e. FluidigmAssay) The constructor
\code{SCASet} should be used to make objects of this class.
} \section{Slots}{ \describe{ \item{set}{A \code{list} of
\code{SingleCellAssays} or its subclasses} }
}

