% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAssay-methods.R
\name{SceToSingleCellAssay}
\alias{SceToSingleCellAssay}
\title{Coerce a SingleCellExperiment to some class defined in MAST}
\usage{
SceToSingleCellAssay(sce, class = "SingleCellAssay",
  check_sanity = TRUE)
}
\arguments{
\item{sce}{object inheriting from \code{SingleCellExperiment}}

\item{class}{\code{character} naming the class to be coerced to}

\item{check_sanity}{(default: \code{TRUE}) Set \code{FALSE} to override sanity checks that try to ensure that the default assay is log-transformed and has at least one exact zero.  See \link{defaultAssay} for details on the "default assay" which is assumed to contain log transformed data.}
}
\value{
object of the indicated class.
}
\description{
Coerce a SingleCellExperiment to some class defined in MAST
}
