% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zlmHooks.R
\name{score_2_hook}
\alias{score_2_hook}
\title{Hook to compute the combined residuals with treatment effects added back.}
\usage{
score_2_hook(x)
}
\arguments{
\item{x}{the ZLMFit}
}
\description{
Hook to compute the combined residuals with treatment effects added back.
}

