% Generated by roxygen2 (4.0.2): do not edit by hand
\name{score_hook}
\alias{score_hook}
\title{Used to compute module "scores"}
\usage{
score_hook(x)
}
\arguments{
\item{x}{the ZLMFit}
}
\description{
Hook to compute y_i-E(V_i)-\beta_{ngeneson}*ng_i for y_i>0 and E(U_i)(E(V_i)-\beta_{ngeneson}*ng_i) for y_i=0
}

