% Generated by roxygen2 (4.0.1): do not edit by hand
\name{se.coef}
\alias{se.coef}
\title{Standard error on coefficients}
\usage{
se.coef(object, ...)
}
\arguments{
\item{object}{ZlmFit}

\item{which}{character vector, one of "C" (continuous) or "D" (discrete) specifying which component should be returned}
}
\value{
matrix of standard errors
}
\description{
Standard error on coefficients
}

