% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shalek_weights}
\alias{shalek_weights}
\title{Hook to return p_hat from the model}
\usage{
shalek_weights(x)
}
\arguments{
\item{x}{the ZLMFit}
}
\description{
Used to weight the unobserved values in a module score calculation analogous to what is done in Shalek et.al.
}

