% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,Mapping-method}
\alias{show,DataLayer-method}
\alias{show,LMlike-method}
\alias{show,Mapping-method}
\alias{show,SCASet-method}
\alias{show,SingleCellAssay-method}
\alias{show,ThresholdedNanoString-method}
\alias{show,ZlmFit-method}
\title{Show methods}
\usage{
\S4method{show}{Mapping}(object)

\S4method{show}{DataLayer}(object)

\S4method{show}{LMlike}(object)

\S4method{show}{ThresholdedNanoString}(object)

\S4method{show}{SCASet}(object)

\S4method{show}{SingleCellAssay}(object)

\S4method{show}{ZlmFit}(object)
}
\arguments{
\item{object}{whose human-readable form is desired.}
}
\description{
Show methods
}
\section{Methods (by class)}{
\itemize{
\item \code{Mapping}: 

\item \code{LMlike}: 

\item \code{ThresholdedNanoString}: 

\item \code{SCASet}: 

\item \code{SingleCellAssay}: 

\item \code{ZlmFit}: 
}}

