% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{simulate1Unit}
\alias{simulate1Unit}
\title{Simulate zero-inflated single cell data}
\usage{
simulate1Unit(x, betaC, betaD, sigmaC)
}
\arguments{
\item{x}{covariate matrix.  number of rows determines number of observations.}

\item{betaC}{continuous coefficient vector}

\item{betaD}{discrete coefficient vector}

\item{sigmaC}{sd of continuous component}
}
\value{
vector of observations y
}
\description{
Simulate zero-inflated single cell data
}

