% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{methods}
\name{SingleCellAssay}
\alias{SingleCellAssay}
\title{SingleCellAssay: A constructor for an object of type SingleCellAssay.}
\usage{
SingleCellAssay(dataframe = NULL, idvars = NULL, primerid = NULL,
  measurement = NULL, id = numeric(0), cellvars = NULL,
  featurevars = NULL, phenovars = NULL, ...)
}
\arguments{
\item{dataframe}{A 'flattened' \code{data.frame} or \code{data.table} containing columns giving cell and feature identifiers and  a measurement column}

\item{idvars}{character vector naming columns that uniquely identify a cell}

\item{primerid}{character vector of length 1 that names the column that identifies what feature (i.e. gene) was measured}

\item{measurement}{character vector of length 1 that names the column containing the measurement}

\item{id}{An identifier (eg, experiment name) for the resulting object}

\item{cellvars}{Character vector naming columns containing additional cellular metadata}

\item{featurevars}{Character vector naming columns containing additional feature metadata}

\item{phenovars}{Character vector naming columns containing additional phenotype metadata}

\item{...}{additional arguments are ignored}
}
\value{
SingleCellAssay object
}
\description{
This is the constructor for the class. This class intends to ease the analysis of single cell assays, in which multiple, exchangeable, cells from an experimental unit (patient, or organism) are assayed along several (or many) dimensions, such as genes. A few examples of this might be Fluidigm gene expression chips, or single cell sequencing experiments.  The chief functionality is to make it easy to keep cellular-level metadata linked to the measurements through \code{cellData} and \code{phenoData}.  There are also subsetting and splitting measures to coerce between a SingleCellAssay, and a \link{SCASet}.
}
\examples{
## See FluidigmAssay for examples
\dontrun{example(FluidigmAssay)}
}
\seealso{
\code{\link{FluidigmAssay}}
}

