\name{SingleCellAssayValidity}
\alias{SingleCellAssayValidity}
\title{SingleCellAssayValidity}
\usage{
  SingleCellAssayValidity(object)
}
\arguments{
  \item{object}{A \code{SingleCellAssay} object or object
  inheriting from SingleCellAssay}
}
\value{
  A \code{logical}. \code{TRUE} if the object is valid,
  \code{FALSE} otherwise.
}
\description{
  Function to check the validity of SingleCellAssay
  objects. More specific functionality can wrap this
  function.
}

