% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{som_admin2}
\alias{som_admin2}
\title{Somalia Admin 2 Names with region and code}
\format{
A data.frame of district, region and code
}
\usage{
som_admin2
}
\description{
A data frame of admin with region, district and code
}
\keyword{datasets}
