% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{som_admins}
\alias{som_admins}
\title{Somalia Admin Names}
\format{
A data.frame of "full" (all names) and "old" (names to use)
}
\usage{
som_admins
}
\description{
A data frame of admin names to match against
}
\keyword{datasets}
