% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{som_shp}
\alias{som_shp}
\title{Somalia Shape File}
\format{
A sf object with "mast_name" giving admin names matching "som_admins"
}
\source{
\url{https://data.humdata.org/dataset/cod-ab-som?}
}
\usage{
som_shp
}
\description{
Shape file for Somalia
}
\keyword{datasets}
