% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLayer.R
\docType{methods}
\name{[,DataLayer-method}
\alias{[,DataLayer-method}
\title{Subset a DataLayer}
\usage{
\S4method{[}{DataLayer}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{DataLayer}

\item{i}{boolean or integer index}

\item{j}{boolean or integer index}

\item{...}{providing an extra index is an error}

\item{drop}{ignored}
}
\value{
DataLayer
}
\description{
Returns a subsetted DataLayer
}

