% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SCASet-methods.R
\docType{methods}
\name{[,SCASet,ANY,ANY-method}
\alias{[,SCASet,ANY,ANY-method}
\title{Subset an SCASet to a smaller SCASet}
\usage{
\S4method{[}{SCASet,ANY,ANY}(x, i, j, ...drop = FALSE)
}
\arguments{
\item{x}{\code{SCASet} to be subscripted}

\item{i}{index}

\item{j}{ignored}

\item{...}{ignored}

\item{drop}{ignored}
}
\value{
An \code{SCASet}
}
\description{
Subset an SCASet to a smaller SCASet
}
\details{
\code{signature(x="SCASet",i="ANY")}: \code{x[i]}, where \code{i} is the vector of integers or characters matching sampleNames. Returns an SCASet of \code{length(i)} of elements at positions in \code{i}.
}

