% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{methods}
\name{[,SingleCellAssay,ANY,ANY,ANY-method}
\alias{[,SingleCellAssay,ANY,ANY,ANY-method}
\alias{[,SingleCellAssay,ANY-method}
\title{Subset a SingleCellAssay}
\usage{
\S4method{[}{SingleCellAssay,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{SingleCellAssay}

\item{i}{logical, integer or character (naming wellKeys)}

\item{j}{logical, integer or character (naming primerid)}

\item{...}{ignored}

\item{drop}{ignored}
}
\value{
SingleCellAssay, suitably subsetted
}
\description{
Subset a SingleCellAssay
}
\details{
\code{signature(x="SingleCellAssay", i="ANY")}: \code{x[i]}, where \code{i} is a logical, integer, or character vector, recycled as necessary to match \code{nrow(x)}. Optional \code{x[[i,j]]} where j is a logical, integer or character vector selecting the features based on ``primerid'' which is unique, while ``geneid'' or gene name is not necessarily unique.
}
\note{
x[[i,j]] functions similarly, but this behavior may change in future releases.
}

