% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SingleCellAssay-methods.R
\docType{methods}
\name{subset,SingleCellAssay-method}
\alias{subset,SingleCellAssay-method}
\title{Subset a SingleCellAssay by data in cellData}
\usage{
\S4method{subset}{SingleCellAssay}(x, thesubset, ...)
}
\arguments{
\item{x}{SingleCellAssay}

\item{thesubset}{expression, which when evaluated in cellData environment which returns a logical}
}
\description{
Subset a SingleCellAssay by data in cellData
}
\examples{
data(vbetaFA)
subset(vbetaFA, ncells==1)
}

