% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SingleCellAssay-methods.R
\docType{methods}
\name{subset}
\alias{subset}
\alias{subset,SingleCellAssay-method}
\title{Subsetting Vectors, Matrices and Data Frames}
\usage{
subset(x, ...)

\S4method{subset}{SingleCellAssay}(x, ...)
}
\arguments{
\item{x}{\code{SingleCellAssay}}

\item{...}{expression}
}
\value{
\code{SingleCellAssay}
}
\description{
Return subsets of vectors, matrices or data frames which meet conditions.

Evaluates the expression in \code{...} in the context of \code{colData(x)} and returns a subsetted version of \code{x}
}

