% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarize}
\alias{summarize}
\title{Return programmatically useful summary of a fit}
\usage{
summarize(object, ...)
}
\arguments{
\item{object}{LMlike or subclass}
}
\value{
list of parameters characterizing fit
}
\description{
Return programmatically useful summary of a fit
}

