% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{summary}
\alias{summary}
\title{Return a human-readable summary}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{some object to be summarized}

\item{...}{additional arguments passed to methods.}
}
\value{
a summary, possibly printed
}
\description{
Return a human-readable summary
}

