% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdSCRNA.R
\name{summary.thresholdSCRNACountMatrix}
\alias{summary.thresholdSCRNACountMatrix}
\alias{print.summaryThresholdSCRNA}
\title{Summarize the effect of thresholding}
\usage{
\method{summary}{thresholdSCRNACountMatrix}(object, ...)

\method{print}{summaryThresholdSCRNA}(x, ...)
}
\arguments{
\item{object}{a \code{thresholdSCRNACountMatrix}}

\item{...}{currently ignored}

\item{x}{a \code{summaryThresholdSCRNA} object, ie output from \code{summary.thresholdSCRNACountMatrix}}
}
\value{
a list of statistics on the original data, and thresholded data
}
\description{
Returns the proportion of (putative) expression, the variance of expressed cells, and -log10 shapiro-wilk tests for normality on the expressed cells
}
\section{Functions}{
\itemize{
\item \code{print(summaryThresholdSCRNA)}: prints five-number distillation of the statistics and invisibly returns the table used to generate the summary

}}
