% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{test.zlm}
\alias{test.zlm}
\title{Likelihood ratio test for hurdle model}
\usage{
test.zlm(model, hypothesis.matrix, type = "Wald", silent = TRUE, ...)
}
\arguments{
\item{model}{output from zlm}

\item{hypothesis.matrix}{argument passed to lht, or string naming a single variable to be dropped from the model, or a formula that is a subset of 'model'}

\item{type}{Test using Wald test or Likelihood Ratio test}

\item{silent}{Silence common errors in testing}
}
\value{
array containing the discrete, continuous and combined tests
}
\description{
Do LR test separately on continuous and discrete portions
Combine for testing hypothesis.matrix
}
\details{
It tests the provided hypothesis.matrix using a Chi-Squared Wald or LRT test
}
\seealso{
zlm
}

