% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ThresholdedNanoString-class}
\alias{ThresholdedNanoString-class}
\title{Holds output and diagnostics from thresholdNanoString
Not intended to be called by the user.}
\description{
Holds output and diagnostics from thresholdNanoString
Not intended to be called by the user.
}
\section{Slots}{

\describe{
\item{melted}{A \code{data.frame} containing a melted version of \code{nsa}, plus the columns 'ps', giving the probability that a measurement belongs to the signal cluster, 'clusterID' the inferred cluster}
\item{nsa}{The thresholded \code{NanoStringAssay} with the thresholded expression in layer \code{et}}
\item{densities}{A \code{list} of length \code{ncol(nsa)} of marginal (mixture model) densities of each gene.}
\item{means}{A \code{matrix} dimension \code{ncol(nsa)} \eqn{\times} 2 given the posterior mean of each cluster.}
\item{props}{A \code{matrix} dimension \code{ncol(nsa)} \eqn{\times} 2 given the posterior probability of each cluster.}
\item{startLayer}{A \code{character} giving the initial layer that was used to generate the thresholding}
}
}
\seealso{
thresholdNanoString
}

