% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/thresholdSCRNA.R
\name{thresholdMMfit}
\alias{thresholdMMfit}
\title{thresholdMMfit}
\usage{
thresholdMMfit(log_data_list = NULL, cutpoints = NULL, plot = FALSE,
  G = 3)
}
\arguments{
\item{log_data_list}{the list of binned data}

\item{cutpoints}{the vector of cutpoints}

\item{plot}{logical whether to show plots or not}

\item{G}{number of components}
}
\description{
Fit a gamma distribution to the data below the cutpoint. Fit a G component mixture to the data above the cutpoint.
Compute the full density and calculate the mixture weights for each observation.
}
\author{
Greg Finak
}

