% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{thresholdNanoString,NanoStringAssay-method}
\alias{thresholdNanoString,NanoStringAssay-method}
\title{Estimate thresholds for positive expression}
\usage{
\S4method{thresholdNanoString}{NanoStringAssay}(nsa, include.primers,
  exclude.primers, posteriorprob, clip = c("left", "right", "NA"),
  debug = FALSE, location.strength = 1, pseudo.counts = 5,
  hard.threshold = 3, startLayer = "lCount")
}
\arguments{
\item{nsa}{NanostringAssay object}

\item{include.primers}{primeris to use for population estimates.  If missing, then all primers are used.}

\item{exclude.primers}{primers to exclude from population estimates}

\item{posteriorprob}{currently ignored}

\item{clip}{currently ignored}

\item{debug}{should extra fields helpful for examining the thresholding be returned as a list?}

\item{location.strength}{scaling of prior on location}

\item{pseudo.counts}{total strength of prior vs data}

\item{hard.threshold}{location estimates less than this value will be treated as belonging to the "noise cluster"}

\item{startLayer}{which layer should be used to threshold.  Default 'lCount'.}
}
\value{
object of class \code{ThresholdedNanoString} if \code{debug=TRUE}, otherwise returns \code{nsa} with the thresholded expression in layer \code{et}
}
\description{
Estimates per-gene x unit thresholds for positive expression and truncates values below this threshold
Uncertain values (in terms of posterior probability of membership) can be set to NA or rounded left or right
Thresholds are estimated using a Gaussian mixture model with prior supplied by population estimates.
}
\seealso{
ThresholdedNanoString-class
}

