% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thresholdSCRNA.R
\name{thresholdRNASeqSingleCellAssay}
\alias{thresholdRNASeqSingleCellAssay}
\title{Threshold and RNASeq SingleCellAssay object}
\usage{
thresholdRNASeqSingleCellAssay(assay = NULL, layername = "count",
  isLog = FALSE, ...)
}
\arguments{
\item{assay}{\code{SingleCellAssay} object.}

\item{layername}{\code{character} the name of the layer to use for thresholding. Defaults to \code{"count"}.}

\item{isLog}{\code{logical} default \code{FALSE}, specifies if \code{layername} is log transformed.}

\item{...}{additional arguments passed on to \code{thresholdSCRNACountMatrix}.
\code{conditions} NULL by default, \code{cutbins} pre-specified cutpoints, \code{nbins} number of bins, \code{bin_by} bin by "median" by default, \code{qt} quantile 0.975 default, \code{min_per_bin} min features per bin (50), \code{absolute_min} lower level below which expression is off, \code{return_log} return a log transformed matrix (TRUE), \code{G} number of mixture components to fit (2), \code{plot} the mixture fit (FALSE), \code{adj} adjust bandwidth (1).}
}
\value{
\code{SingleCellAssay} with additional layers \code{layername_t} the thresholded layer, and \code{weights} the estimated weights for each observation.
}
\description{
Threshold SingleCellAssay
}
\details{
Threshold a SingleCellAssay object. Uses the "count" layer by default to call \code{thresholdSCRNACountMatrix}. The parameter \code{isLog} is a \code{logical} specifying whether \code{layername} is log transformed (i.e. log counts).
}

