% Generated by roxygen2 (4.0.2): do not edit by hand
\name{thresholdSCRNACountMatrix}
\alias{thresholdSCRNACountMatrix}
\title{Threshold a count matrix using an adaptive threshold.}
\usage{
thresholdSCRNACountMatrix(data_all, conditions = NULL, cutbins = NULL,
  nbins = 10, bin_by = "median", qt = 0.975, min_per_bin = 50,
  absolute_min = 1, log_base = 2)
}
\arguments{
\item{data_all}{\code{matrix} of counts}

\item{conditions}{??}

\item{cutbins}{\code{vector} of cut points.}

\item{nbins}{\code{integer} number of bins when cutbins is not specified.}

\item{bin_by}{\code{character} "median", "proportion", "mean"}

\item{qt}{??}

\item{min_per_bin}{??}

\item{absolute_min}{??}

\item{log_base}{??}

\item{plot}{\code{logical}.}
}
\value{
\code{list} of thresholded counts, thresholds, and bins
}
\description{
An adaptive threshold is calculated from the conditional mean of expression, based on 10 bins
of the genes with similar expression levels. Thresholds are chosen by estimating cutpoints in the bimodal density estimates of the
binned data.
}

