% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{vbeta}
\alias{vbeta}
\title{Vbeta Data Set}
\format{a data frame with 11 columns.
Column \code{Ct} contains the cycle threshold, with NA denoting that the threshold was never crossed.  So it is inversely proportional to the log2 mRNA, and should be negated (and NAs set to zero) if it is used as a expression measurement for a \code{FluidigmAssay}.}
\description{
Vbeta Data Set
}
