% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{vbetaFA}
\alias{vbetaFA}
\title{Vbeta Data Set, FluidigmAssay}
\format{a \code{FluidigmAssay} of the vbeta data set.}
\description{
Vbeta Data Set, FluidigmAssay
}
\seealso{
\code{\link{vbeta}}, \code{\link{FluidigmAssay}}
}

