% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZlmFit.R
\docType{methods}
\name{waldTest,ZlmFit,matrix-method}
\alias{waldTest,ZlmFit,matrix-method}
\title{Wald test}
\usage{
\S4method{waldTest}{ZlmFit,matrix}(object, hypothesis)
}
\arguments{
\item{object}{ZlmFit}

\item{hypothesis}{See Details}
}
\value{
3D array
}
\description{
A 3D array with first dimension being the genes,
next dimension giving information about the test
(the degrees of freedom, Chisq statistic, and P value), and final dimension
being the value of these quantities on the
discrete, continuous and hurdle (combined) levels.
}
