% Generated by roxygen2 (4.0.1): do not edit by hand
\name{waldTest}
\alias{waldTest}
\title{Run a Wald test}
\usage{
waldTest(object, hypothesis)
}
\arguments{
\item{object}{LMlike or subclass}

\item{hypothesis.matrix}{argument suitable to be passed to car::lht}
}
\value{
array giving test statistics
}
\description{
Run a Wald tests on discrete and continuous components
}
\seealso{
fit

lrTest

lht
}

