% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{xform}
\alias{xform}
\title{Make matrix of continuous expression values, orthogonal to discrete}
\usage{
xform(mat, scale = FALSE)
}
\arguments{
\item{mat}{matrix (such as produced by exprs)}

\item{scale}{should the columns also be scaled to have unit variance}
}
\description{
This centers each column of \code{mat} around the mean of its non-zero values.
}

