% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ZlmFit-class}
\alias{ZlmFit-class}
\alias{coef,ZlmFit-method}
\alias{lrTest,ZlmFit,CoefficientHypothesis-method}
\alias{lrTest,ZlmFit,Hypothesis-method}
\alias{lrTest,ZlmFit,matrix-method}
\alias{se.coef,ZlmFit-method}
\alias{summary,ZlmFit-method}
\alias{vcov,ZlmFit-method}
\alias{waldTest,ZlmFit,CoefficientHypothesis-method}
\alias{waldTest,ZlmFit,Hypothesis-method}
\title{An S4 class to hold the output of a call to zlm}
\usage{
\S4method{lrTest}{ZlmFit,CoefficientHypothesis}(object, hypothesis)

\S4method{lrTest}{ZlmFit,Hypothesis}(object, hypothesis)

\S4method{lrTest}{ZlmFit,matrix}(object, hypothesis)

\S4method{waldTest}{ZlmFit,CoefficientHypothesis}(object, hypothesis)

\S4method{waldTest}{ZlmFit,Hypothesis}(object, hypothesis)

\S4method{coef}{ZlmFit}(object, which, ...)

\S4method{vcov}{ZlmFit}(object, which, ...)

\S4method{se.coef}{ZlmFit}(object, which, ...)

\S4method{summary}{ZlmFit}(object, logFC = TRUE, doLRT = FALSE, ...)
}
\arguments{
\item{object}{ZlmFit}

\item{hypothesis}{call to \link{Hypothesis} or \link{CoefficientHypothesis} or a matrix giving such contrasts.}

\item{which}{character vector, one of "C" (continuous) or "D" (discrete) specifying which component should be returned}

\item{...}{ignored}

\item{logFC}{If TRUE, calculate log-fold changes, or output from a call to \code{getLogFC}.}

\item{doLRT}{if TRUE, calculate lrTests on each coefficient, or a character vector of such coefficients to consider.}
}
\description{
This holds output from a call to zlm.SingleCellAssay.  Many methods are defined to operate on it.  See below.


}
\section{Methods (by generic)}{
\itemize{
\item \code{lrTest}: Returns an array with likelihood-ratio tests on contrasts defined using \code{CoefficientHypothesis()}.

\item \code{lrTest}: Returns an array with likelihood-ratio tests specified by \code{Hypothesis}, which is a \link{Hypothesis}.

\item \code{lrTest}: Returns an array with likelihood-ratio tests specified by \code{Hypothesis}, which is a contrast \code{matrix}.

\item \code{waldTest}: Returns an array with Wald Tests on contrasts defined using \code{CoefficientHypothesis()}.

\item \code{waldTest}: Returns an array with Wald Tests on contrasts defined in \code{Hypothesis()}

\item \code{coef}: Returns the matrix of coefficients for component \code{which}.

\item \code{vcov}: Returns an array of variance/covariance matrices for component \code{which}.

\item \code{se.coef}: Returns a matrix of standard error estimates for coefficients on component \code{which}.

\item \code{summary}: Returns a \code{data.table} summary of fit (invisibly).
}}
\section{Slots}{

\describe{
\item{\code{coefC}}{}

\item{\code{coefD}}{matrices of coefficients}

\item{\code{vcovC}}{}

\item{\code{vcovD}}{array of variance/covariance matrices for coefficients}

\item{\code{LMlike}}{the LmWrapper object used}

\item{\code{sca}}{the \code{SingleCellAssay} object used}

\item{\code{deviance}}{}

\item{\code{loglik}}{}

\item{\code{df.null}}{}

\item{\code{df.resid}}{}

\item{\code{dispersion}}{}

\item{\code{dispersionNoShrink}}{}

\item{\code{priorDOF}}{}

\item{\code{priorVar}}{}

\item{\code{converged}}{output that may optionally be set by the underlying modeling function}

\item{\code{hookOut}}{a list of length ngenes containing output from a hook function, if \code{zlm} was called with one}
}}
\seealso{
zlm.SingleCellAssay
}

