% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{addlayer}
\alias{addlayer}
\title{Layer management methods}
\usage{
addlayer(x, name)
}
\arguments{
\item{x}{a \code{SingleCellAssay} or derived}

\item{name}{a \code{code} character}
}
\description{
A layer is a "slice" of a 3d array.  \code{SingleCellAssay} objects are stateful, in that a layer can be set, and then various routines will operate on that layer.
}
\section{Methods}{

}
\examples{
data(vbetaFA)
vbetaFA <- addlayer(vbetaFA, 'expressed')
layername(vbetaFA)
ee <- exprs(vbetaFA)
layer(vbetaFA) <- 'expressed'
exprs(vbetaFA) <- (ee>0)*1
layer(vbetaFA) <- 'Et'
layername(vbetaFA) <- 'et'
layername(vbetaFA)
nlayer(vbetaFA)
}

