% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZlmFit-bootstrap.R
\name{bootVcov1}
\alias{bootVcov1}
\title{Bootstrap a zlmfit}
\usage{
bootVcov1(zlmfit, R = 99)
}
\arguments{
\item{zlmfit}{class \code{ZlmFit}}

\item{R}{number of bootstrap replicates}
}
\value{
array of bootstrapped coefficients
}
\description{
Sample cells with replacement to find bootstrapped distribution of coefficients
}
\examples{
data(vbetaFA)
zlmVbeta <- zlm(~ Stim.Condition, subset(vbetaFA, ncells==1)[1:5,])
#Only run 3 boot straps, which you wouldn't ever want to do in practice...
bootVcov1(zlmVbeta, R=3)
}
