% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GSEA-by-boot.R
\name{calcZ}
\alias{calcZ}
\title{Get Z or T statistics and P values after running gseaAfterBoot}
\usage{
calcZ(tests, testType = "t")
}
\arguments{
\item{tests}{output from \code{gseaAfterBoot}}

\item{testType}{either 'normal' or 't'.  The 't' test adjusts for excess kurtosis due to the finite number of bootstrap replicates used to estimate the variance of the statistics.  This will result in more conservative inference.}
}
\value{
3D array with dimensions set (modules) comp ('cont'inuous or 'disc'rete) and metric ('Z' stat and two sided 'P' value that P(z>|Z|))
}
\description{
Get Z or T statistics and P values after running gseaAfterBoot
}
\seealso{
gseaAfterBoot
}

