% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcZ}
\alias{calcZ}
\title{Get Z statistics and P values after running gseaAfterBoot}
\usage{
calcZ(tests)
}
\arguments{
\item{tests}{output from \code{gseaAfterBoot}}
}
\value{
3D array with dimensions set (modules) comp ('cont'inuous or 'disc'rete) and metric ('Z' stat or two sided 'P' value that P(z>|Z|))
}
\description{
Get Z statistics and P values after running gseaAfterBoot
}

