% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{methods}
\name{cData}
\alias{cData}
\title{cData}
\usage{
cData(sc)
}
\arguments{
\item{sc}{An object with \code{cellData}}
}
\value{
\code{data.frame}
}
\description{
Accessor for cellData \code{data.frame}
}
\details{
Returns the \code{cellData} \code{data.frame}.
}
\keyword{accessor}

