% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SingleCellAssay-methods.R
\docType{methods}
\name{cData,SingleCellAssay-method}
\alias{cData,SingleCellAssay-method}
\title{Get cellData data.frame}
\usage{
\S4method{cData}{SingleCellAssay}(sc)
}
\arguments{
\item{sc}{SingleCellAssay}
}
\value{
data.frame
}
\description{
Get cellData data.frame
}

