% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SingleCellAssay-methods.R
\docType{methods}
\name{cData}
\alias{cData}
\alias{cData<-}
\alias{cData<-,SingleCellAssay-method}
\alias{combine,SingleCellAssay,ANY-method}
\alias{combine,SingleCellAssay,SingleCellAssay-method}
\title{Accessor for cellData \code{data.frame}}
\usage{
cData(sc)

cData(sc) <- value

\S4method{cData}{SingleCellAssay}(sc) <- value

\S4method{combine}{SingleCellAssay,SingleCellAssay}(x, y, ...)

\S4method{combine}{SingleCellAssay,ANY}(x, y, ...)
}
\arguments{
\item{sc}{An object with \code{cellData}}

\item{value}{Replacement, one of \code{AnnotatedDataFrame} or \code{data.frame}}
}
\value{
\code{data.frame}
}
\description{
Accessor for cellData \code{data.frame}
}
\details{
\code{cData(sc)}: Return the \code{cellData} \code{data.frame}.

\code{cData(sc)<-value}: Replace the cellData with \code{value}, which can be either an \code{AnnotatedDataFrame} or \code{data.frame}.  The replacement is checked that it has mandatory fields defined by its class.
}
\section{Methods (by class)}{
\itemize{
\item \code{SingleCellAssay}: replace the \code{colData}

\item \code{x = SingleCellAssay,y = SingleCellAssay}: combine two experiments along rows/columns

\item \code{x = SingleCellAssay,y = ANY}: combine two experiments along rows/columns
}}

