% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SingleCellAssay-methods.R
\docType{methods}
\name{cData}
\alias{cData}
\alias{cData<-}
\alias{cData,SingleCellAssay-method}
\alias{cData<-,SingleCellAssay-method}
\alias{combine,SingleCellAssay,SingleCellAssay-method}
\alias{combine,SingleCellAssay,ANY-method}
\title{Deprecated cell/feature data accessors/mutators}
\usage{
cData(sc)

cData(sc) <- value

\S4method{cData}{SingleCellAssay}(sc)

\S4method{cData}{SingleCellAssay}(sc) <- value

\S4method{combine}{SingleCellAssay,SingleCellAssay}(x, y, ...)

\S4method{combine}{SingleCellAssay,ANY}(x, y, ...)
}
\arguments{
\item{sc}{An object with \code{cellData}}

\item{value}{replacement value}

\item{x}{\code{SingleCellAssay}}

\item{y}{\code{SingleCellAssay}}

\item{...}{\code{SingleCellAssay}}
}
\value{
\code{DataFrame} or modifies the \code{SingleCellAssay} object in place
}
\description{
These functions are now all deprecated and will be removed in a future release.
}
\details{
\code{cData(sc)}: Return the \code{cellData} \code{data.frame}.

\code{cData(sc)<-value}: Replace the cellData with \code{value}, which can be either an \code{AnnotatedDataFrame} or \code{data.frame}.  The replacement is checked that it has mandatory fields defined by its class.

\code{combine(x, y, ...)}: Concatenate two experiments along rows/columns
}
\section{Replacement Functions}{

You should transition to use the following replacements:
\describe{
  \item{\code{cData}}{\link{colData}}
  \item{\code{fData}}{\link{mcols}}
  \item{\code{exprs}}{\link{assay}}
  \item{\code{combine}}{\link{cbind2} or \link{rbind2}}
}
}

\examples{
data(vbetaFA)
stopifnot(all.equal(hushWarning(cData(vbetaFA), 'deprecated'),colData(vbetaFA)))
stopifnot(all.equal(hushWarning(fData(vbetaFA), 'deprecated'), mcols(vbetaFA)))
stopifnot(all.equal(hushWarning(exprs(vbetaFA), 'deprecated'), t(assay(vbetaFA))))
}
\seealso{
exprs
}
