% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coef}
\alias{coef}
\alias{coef,LMlike-method}
\title{Coefficients of zero-inflated}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{LMlike}

\item{...}{passed to methods}
}
\value{
numeric vector
}
\description{
Given a fitted LMlike, return the coefficients from discrete or continuous.
Methods expect an argument \code{which}, a \code{character} of length one, one of "C" (continuous) or "D" (discrete) specifying which component should be returned.
}

