% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{conform}
\alias{conform}
\title{Do two objects conform in dimension and type?}
\usage{
conform(dl, other)
}
\arguments{
\item{dl}{DataLayer}

\item{other}{Another object}
}
\value{
bitmask containing number of dimensions that agree
}
\description{
Returns false if:
other does not inherit from Matrix
Otherwise returns a bitmask, which is the sum of:
1: objects have same number of rows
2: objects have same number of columns
4: objects have same number of layers
}
\author{
andrew
}

