% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zlmHooks.R
\name{deviance_residuals_hook}
\alias{deviance_residuals_hook}
\title{Standardized deviance residuals hook}
\usage{
deviance_residuals_hook(x)
}
\arguments{
\item{x}{the ZLMFit}
}
\description{
Computes the average of the standardized deviance residuals for the discrete and continuous models
}

