% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zlmHooks.R
\name{discrete_residuals_hook}
\alias{discrete_residuals_hook}
\title{hHook to get the continuous residuals}
\usage{
discrete_residuals_hook(x)
}
\arguments{
\item{x}{the ZLMFit}
}
\description{
hHook to get the continuous residuals
}

