% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fluidigm-methods.R
\name{expavg}
\alias{expavg}
\title{Exponential average}
\usage{
expavg(x)
}
\arguments{
\item{x}{\code{numeric}}
}
\value{
\code{numeric}
}
\description{
Puts log transformed values onto natural scale and takes mean of vector.
Calculates mean(2^x - 1)
}
\examples{
x <- 1:10
logmean(expavg(x))
}
