% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLayer.R
\docType{methods}
\name{exprs,DataLayer-method}
\alias{exprs,DataLayer-method}
\alias{exprs<-,DataLayer,ANY-method}
\title{Get or set a matrix of measurement values in a \code{SingleCellAssay}}
\usage{
\S4method{exprs}{DataLayer}(object)

\S4method{exprs}{DataLayer,ANY}(object) <- value
}
\arguments{
\item{object}{\code{SingleCellAssay} or inheriting class}
}
\value{
a \code{matrix} of measurement values with wells on the rows and features on the columns of the default layer
}
\description{
Return or set a matrix of the measurement: cells by primerids.  \strong{Note this is the transpose of ExpressionSets}.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = DataLayer,value = ANY}: 
}}

