% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/DataLayer.R
\docType{methods}
\name{exprsLayer}
\alias{exprs,DataLayer-method}
\alias{exprs<-,DataLayer,ANY-method}
\alias{exprsLayer}
\alias{exprsLayer,DataLayer,character-method}
\alias{exprsLayer,DataLayer,missing-method}
\alias{exprsLayer,DataLayer,numeric-method}
\alias{exprsLayer<-}
\alias{exprsLayer<-,DataLayer,character-method}
\alias{exprsLayer<-,DataLayer,missing-method}
\alias{exprsLayer<-,DataLayer,numeric-method}
\title{Get or set expression matrix by layername}
\usage{
exprsLayer(object, layernm, ...)

\S4method{exprsLayer}{DataLayer,numeric}(object, layernm)

\S4method{exprsLayer}{DataLayer,character}(object, layernm)

\S4method{exprsLayer}{DataLayer,missing}(object, layernm)

\S4method{exprs}{DataLayer}(object)

\S4method{exprsLayer}{DataLayer,numeric}(object, layernm, ...) <- value

\S4method{exprsLayer}{DataLayer,character}(object, layernm, ...) <- value

\S4method{exprsLayer}{DataLayer,missing}(object, layernm, ...) <- value

\S4method{exprs}{DataLayer,ANY}(object) <- value
}
\arguments{
\item{object}{DataLayer or subclass.}

\item{layernm}{character, numeric or missing.}

\item{value}{replacement value with conforming dimensions as layer \code{layernm}.}
}
\value{
a \code{matrix} of measurement values with wells on the rows and features on the columns of the default layer.
}
\description{
Get or set expression matrix by layername
}
\section{Methods (by class)}{
\itemize{
\item \code{object = DataLayer,layernm = numeric}: Return layer \code{layernm}.

\item \code{object = DataLayer,layernm = character}: Return layer \code{layernm}.

\item \code{object = DataLayer,layernm = missing}: Return current layer.

\item \code{DataLayer}: Return current layer.

\item \code{object = DataLayer,layernm = numeric}: Replace layer \code{layernm}.

\item \code{object = DataLayer,layernm = character}: Replace layer \code{layernm}.

\item \code{object = DataLayer,layernm = missing}: Replace current layer.

\item \code{object = DataLayer,value = ANY}: Replace current layer.
}}
\examples{
data(vbetaFA)
show(vbetaFA)
exprs(vbetaFA)[1:5,1:5]
exprsLayer(vbetaFA, 'Et')[1:5,1:5]
exprsLayer(vbetaFA, 'Et') <- 999
exprs(vbetaFA)[1:5, 1:5]
}

