% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SingleCellAssay-methods.R
\docType{methods}
\name{fData}
\alias{fData}
\alias{fData,SingleCellAssay-method}
\title{fData}
\usage{
\S4method{fData}{SingleCellAssay}(object)
}
\arguments{
\item{object}{An object with \code{featureData}}
}
\value{
\code{data.frame}
}
\description{
Accessor for featureData \code{data.frame}
}
\details{
Returns the \code{featureData} \code{data.frame}.
}
\keyword{accessor}

