\docType{methods}
\name{fData}
\alias{fData}
\alias{fData,SingleCellAssay-method}
\title{fData}
\arguments{
  \item{object}{An object with \code{featureData}}
}
\value{
  \code{data.frame}
}
\description{
  Accessor for featureData \code{data.frame}
}
\details{
  Returns the \code{featureData} \code{data.frame}.
}
\keyword{accessor}

