% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{fData}
\alias{fData}
\alias{fData,SingleCellAssay-method}
\title{fData}
\arguments{
\item{object}{An object with \code{featureData}}
}
\value{
\code{data.frame}
}
\description{
Accessor for featureData \code{data.frame}
}
\details{
Returns the \code{featureData} \code{data.frame}.
}
