\docType{methods}
\name{featureData}
\alias{featureData}
\alias{featureData,SingleCellAssay-method}
\title{Accessor for featureData \code{AnnotatedDataFrame}}
\arguments{
  \item{sc}{An object with \code{featureData}}
}
\value{
  \code{AnnotatedDataFrame}
}
\description{
  Returns the \code{featureData}.
}
\keyword{accessor}

