% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{featureData}
\alias{featureData}
\alias{featureData,SingleCellAssay-method}
\title{Accessor for featureData \code{AnnotatedDataFrame}}
\usage{
\S4method{featureData}{SingleCellAssay}(object)
}
\arguments{
\item{object}{An object with \code{featureData}}
}
\value{
\code{AnnotatedDataFrame}
}
\description{
Returns the \code{featureData}.
}
\keyword{accessor}

