% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{featureData}
\alias{featureData}
\alias{featureData,SingleCellAssay-method}
\title{Accessor for featureData \code{AnnotatedDataFrame}}
\arguments{
\item{object}{An object with \code{featureData}}
}
\value{
\code{AnnotatedDataFrame}
}
\description{
Returns the \code{featureData}.
}
