% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lmWrapper-glmer.R
\docType{methods}
\name{fit,LMERlike,missing-method}
\alias{fit,LMERlike,missing-method}
\title{##'}
\usage{
\S4method{fit}{LMERlike,missing}(object, response, silent = TRUE, ...)
}
\description{
##'
}

