% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/lmWrapper-glmer.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,GLMlike,missing-method}
\alias{fit,LMERlike,missing-method}
\title{fit a zero-inflated regression}
\usage{
fit(object, response, ...)

\S4method{fit}{LMERlike,missing}(object, response, silent = TRUE, ...)
}
\arguments{
\item{object}{inheriting from \code{LMlike}}

\item{response}{a vector, same length as the design, or if missing then use the current response}

\item{...}{currently ignored}

\item{silent}{mute some warnings emitted from the underlying modeling functions}
}
\value{
LMlike or subclass
}
\description{
Given a design and formula, fit the zero inflated regression, storing the fits in slots
\code{fitC} and \code{fitD}
}
