% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{FluidigmAssay}
\alias{FluidigmAssay}
\title{Constructor for a FluidigmAssay}
\usage{
FluidigmAssay(dataframe = NULL, idvars, primerid, measurement, ncells,
  geneid = NULL, id = numeric(0), cellvars = NULL, featurevars = NULL,
  phenovars = NULL, ...)
}
\arguments{
\item{dataframe}{A 'flattened' \code{data.frame} or \code{data.table} containing columns giving cell and feature identifiers and  a measurement column}

\item{idvars}{character vector naming columns that uniquely identify a cell}

\item{primerid}{character vector of length 1 that names the column that identifies what feature (i.e. gene) was measured}

\item{measurement}{character vector of length 1 that names the column containing the measurement}

\item{ncells}{A \code{character} specifying the column which gives the number of cells per well}

\item{geneid}{An optional \code{character} alternate id for primers.}

\item{id}{An identifier (eg, experiment name) for the resulting object}

\item{cellvars}{Character vector naming columns containing additional cellular metadata}

\item{featurevars}{Character vector naming columns containing additional feature metadata}

\item{phenovars}{Character vector naming columns containing additional phenotype metadata}

\item{...}{additional arguments are ignored}
}
\value{
A FluidigmAssay object
}
\description{
Constructs a FluidigmAssay object. Differs little from the SingleCellAssay constructor. Only the \code{ncells} parameter is additionally required.
}
\examples{
data(vbeta)
colnames(vbeta)
vbeta <- computeEtFromCt(vbeta)
vbeta.fa <- FluidigmAssay(vbeta, idvars=c("Subject.ID", "Chip.Number", "Well"),
primerid='Gene', measurement='Et', ncells='Number.of.Cells',
geneid="Gene",cellvars=c('Number.of.Cells', 'Population'),
phenovars=c('Stim.Condition','Time'), id='vbeta all')
show(vbeta.fa)
nrow(vbeta.fa)
ncol(vbeta.fa)
head(fData(vbeta.fa)$primerid)
table(cData(vbeta.fa)$Subject.ID)
vbeta.sub <- subset(vbeta.fa, Subject.ID=='Sub01')
show(vbeta.sub)
}
\author{
Andrew McDavid and Greg Finak
}

