% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{FromMatrix}
\alias{FromMatrix}
\title{Construct a SingleCellAssay from a matrix or array of expression}
\usage{
FromMatrix(class, exprsArray, cData, fData)
}
\arguments{
\item{class}{What class of object are we constructing?}

\item{exprsArray}{matrix or array, rows are cells, columns are genes}

\item{cData}{cellData data.frame or AnnotatedDataFrame}

\item{fData}{featureData data.frame or AnnotatedDataFrame}
}
\value{
an object of class \code{class}
}
\description{
If the gene expression measurements are already in a rectangular form,
then this function allows an easy way to construct a SingleCellAssay object while
still doing some sanity checking of inputs.
}
\examples{
ncells <- 10
ngenes <- 5
fData <- data.frame(primerid=LETTERS[1:ngenes])
cData <- data.frame(wellKey=seq_len(ncells))
mat <- matrix(rnorm(ncells*ngenes), ncol=ngenes)
sca <- FromMatrix('SingleCellAssay', mat, cData, fData)
stopifnot(inherits(sca, 'SingleCellAssay'))
##If there are mandatory keywords expected by a class, you'll have to manually set them yourself
cData$ncells <- 1
fd <- FromMatrix('FluidigmAssay', mat, cData, fData)
stopifnot(inherits(fd, 'FluidigmAssay'))
}

