% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{getLambda}
\alias{getLambda}
\title{Simulate tests on a single cell gene expression data set using hurdle model}
\usage{
getLambda(nsim, thetafun, Xframe, XFormula, Formula, contrasts,
  zlmargs = NULL)
}
\arguments{
\item{nsim}{number of simulations to run}

\item{thetafun}{function returning a realization of the parameters (return a constant if parameters are fixed)}

\item{Xframe}{function returning a `design matrix`, (if the design is random), or a fixed design matrix.}

\item{XFormula}{formula used to translate the design into a model matrix. See details.}

\item{Formula}{formula used to model the design}

\item{contrasts}{argument passed to `lrTest`}

\item{zlmargs}{arguments passed to `zlm.SingleCellAssay`}
}
\value{
matrix containing likelihood ratio stats, p values and dof
}
\description{
Simulate tests on a single cell gene expression data set using hurdle model
}
\examples{
Pclass <- .5
Nind <- 3
Ncell <- 96
Xframe <- data.frame(ind=factor(rep(seq_len(Nind), each=Ncell)), hiv=rbinom(Nind*Ncell, 1, Pclass))
betaD <- betaC <- c('(Intercept)'=10, ind2=-.5, ind3=.5, hiv=.5)
betaD['(Intercept)'] <- 0
theta <-  list(C=betaC, D=betaD, sigmaC=1)
getLambda(10, theta, Xframe, XFormula=~hiv+ind, ~hiv, 'hiv', zlmargs=list(method='bayesglm'))
}

