% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getrc}
\alias{getrc}
\title{getrc}
\usage{
getrc(concord)
}
\arguments{
\item{concord}{data.frame returned by getConcordance}
}
\value{
numeric
}
\description{
Concordance correlation coefficient lin 1989
}
\details{
Return's Lin's intraclass correlation coefficient for concordance
}

