\docType{methods}
\name{getwellKey-method}
\alias{getwellKey}
\alias{getwellKey-method}
\alias{getwellKey,SingleCellAssay-method}
\title{getwellKey}
\arguments{
  \item{sc}{An object with a \code{wellKey}}
}
\value{
  \code{integer} giving the unique id generated
}
\description{
  Accessor for wellKey
}
\details{
  This returns the wellKey, which is a unique identifier
  generated by \code{idvars} in the \code{mapping}
}
\keyword{accessor}

