\name{getwellKey}
\alias{getwellKey}
\alias{getwellKey,SingleCellAssay-method}
\title{Accessor for wellKey}
\usage{
getwellKey(sc)
}
\arguments{
  \item{sc}{An object with a \code{wellKey}}
}
\value{
\code{integer} giving the unique id generated
}
\description{
This returns the wellKey, which is a unique identifier
generated by \code{idvars} in the \code{mapping}
}

