% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getwellKey}
\alias{getwellKey}
\alias{getwellKey,SingleCellAssay-method}
\title{Accessor for wellKey}
\usage{
getwellKey(sc)
}
\arguments{
\item{sc}{An object with a \code{wellKey}}
}
\value{
\code{integer} giving the unique id generated
}
\description{
This returns the wellKey, which is a unique identifier generated by \code{idvars} in the \code{mapping}
}
\examples{
data(vbetaFA)
getwellKey(vbetaFA)
colData(vbetaFA)$wellKey
}
