% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/lmWrapper-glm.R
\docType{class}
\name{GLMlike-class}
\alias{GLMlike-class}
\alias{vcov,GLMlike-method}
\title{Wrapper for regular glm/lm}
\usage{
\S4method{vcov}{GLMlike}(object, which, ...)
}
\arguments{
\item{object}{\code{GLMlike}}

\item{which}{\code{character}, one of 'C', 'D'.}

\item{...}{ignored}
}
\value{
covariance matrix
}
\description{
Wrapper for regular glm/lm
}
\section{Methods (by generic)}{
\itemize{
\item \code{vcov}: return the variance/covariance of component \code{which}
}}

\section{Slots}{

\describe{
\item{\code{weightFun}}{function to map expression values to probabilities of expression.  Currently unused.}
}}

