% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper-methods.R
\name{hushWarning}
\alias{hushWarning}
\title{Selectively muffle warnings based on output}
\usage{
hushWarning(expr, regexp)
}
\arguments{
\item{expr}{an expression}

\item{regexp}{a regexp to be matched (with str_detect)}
}
\value{
the result of expr
}
\description{
Selectively muffle warnings based on output
}

