% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Hypothesis}
\alias{Hypothesis}
\alias{SimpleHypothesis}
\alias{TermHypothesis}
\title{Describe a linear model hypothesis to be tested}
\usage{
Hypothesis(hypothesis)
}
\arguments{
\item{hypothesis}{a character vector specifying a hypothesis, following the format of car::linearHypothesis}
}
\value{
a function to be called with argument 'terms' giving the actual model terms
}
\description{
A \code{Hypothesis} can be any linear combination of coefficients, compared to an arbitrary constant.
A \code{SimpleHypothesis} is a hypothesis for which terms are singly or jointly tested to be zero (generally the case in a t-test or F-test).
A \code{TermHypothesis} is a \code{SimpleHypothesis} for which entire factors are tested jointly.
}
\seealso{
zlm.SingleCellAssay waldTest lrTest linearHypothesis
}

