% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Hypothesis}
\alias{CoefficientHypothesis}
\alias{Hypothesis}
\title{Describe a linear model hypothesis to be tested}
\usage{
Hypothesis(hypothesis)
}
\arguments{
\item{hypothesis}{a character vector specifying a hypothesis, following makeContrasts, or a character vector naming coefficients to be dropped.}
}
\value{
a Hypothesis with a "transformed" component
}
\description{
A \code{Hypothesis} can be any linear combination of coefficients, compared to zero.  Specify it as a character vector that can be parsed to yield the desired equalities ala \code{makeContrasts}.
A \code{CoefficientHypothesis} is a hypothesis for which terms are singly or jointly tested to be zero (generally the case in a t-test or F-test), by dropping coefficients from the model.
}
\seealso{
zlm.SingleCellAssay waldTest lrTest linearHypothesis
}

