% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{layer<-}
\alias{layer<-}
\title{Set active layer}
\usage{
layer(x) <- value
}
\arguments{
\item{x}{DataLayer}

\item{value}{identifier of layer}
}
\value{
DataLayer with new active layer
}
\description{
Set active layer
}
\seealso{
addlayer
}

