% Generated by roxygen2 (4.0.1): do not edit by hand
\name{layer<-}
\alias{layer<-}
\alias{layer<-,DataLayer,numeric-method}
\title{Set active layer}
\usage{
layer(x) <- value
}
\arguments{
\item{x}{DataLayer}

\item{value}{identifier of layer}
}
\value{
DataLayer with new active layer
}
\description{
Set active layer
}

