% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{layername}
\alias{layername}
\title{Get name of active layer}
\usage{
layername(x)
}
\arguments{
\item{x}{DataLayer}
}
\value{
character
}
\description{
Get name of active layer
}
\seealso{
addlayer
}

