% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SCASet-methods.R
\docType{methods}
\name{length,SCASet-method}
\alias{length,SCASet-method}
\title{Get the length of an SCASet}
\usage{
\S4method{length}{SCASet}(x)
}
\arguments{
\item{x}{SCASet}
}
\value{
numeric
}
\description{
Get the length of an SCASet
}

