% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{LMlike-class}
\alias{LMlike-class}
\alias{lrTest,LMlike,CoefficientHypothesis-method}
\alias{lrTest,LMlike,Hypothesis-method}
\alias{lrTest,LMlike,character-method}
\alias{lrTest,LMlike,matrix-method}
\alias{summary,LMlike-method}
\alias{waldTest,LMlike,CoefficientHypothesis-method}
\alias{waldTest,LMlike,matrix-method}
\title{Linear Model-like Class}
\usage{
\S4method{summary}{LMlike}(object)

\S4method{waldTest}{LMlike,CoefficientHypothesis}(object, hypothesis)

\S4method{waldTest}{LMlike,matrix}(object, hypothesis)

\S4method{lrTest}{LMlike,character}(object, hypothesis)

\S4method{lrTest}{LMlike,CoefficientHypothesis}(object, hypothesis)

\S4method{lrTest}{LMlike,Hypothesis}(object, hypothesis)

\S4method{lrTest}{LMlike,matrix}(object, hypothesis)
}
\arguments{
\item{object}{\code{LMlike}}

\item{hypothesis}{one of a \code{CoefficientHypothesis}, \code{Hypothesis} or contrast \code{matrix}.}
}
\description{
Wrapper around modeling function to make them behave enough alike that Wald tests and Likelihood ratio are easy to do.
To implement a new type of zero-inflated model, extend this class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Print a summary of the coefficients in each component.

\item \code{waldTest}: Wald test dropping single term specified by \code{CoefficientHypothesis} \code{hypothesis}

\item \code{waldTest}: Wald test of contrast specified by contrast matrix \code{hypothesis}

\item \code{lrTest}: Likelihood ratio test dropping entire term specified by \code{character} \code{hypothesis} naming a term in the symbolic formula.

\item \code{lrTest}: Likelihood ratio test dropping single term specified by \code{CoefficientHypothesis} \code{hypothesis}

\item \code{lrTest}: Likelihood ratio test dropping single term specified by \code{Hypothesis} \code{hypothesis}

\item \code{lrTest}: Likelihood ratio test dropping single term specified by contrast matrix \code{hypothesis}
}}
\section{Slots}{

\describe{
\item{design}{a data.frame from which variables are taken for the right hand side of the regression}
\item{fitC}{The continuous fit}
\item{fitD}{The discrete fit}
\item{response}{The left hand side of the regression}
\item{fitted}{A \code{logical} with components "C" and "D", TRUE if the respective component has converge}
\item{formula}{A \code{formula} for the regression}
\item{fitArgsC}{}
\item{fitArgsD}{Both \code{list}s giving arguments that will be passed to the fitter (such as convergence criteria or case weights)}
}
}
\seealso{
coef

lrTest

waldTest

vcov

logLik
}

