% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ZlmFit-logFC.R
\name{logFC}
\alias{getLogFC}
\alias{logFC}
\title{Calculate log-fold changes from hurdle model components}
\usage{
logFC(zlmfit, contrast0, contrast1)

getLogFC(zlmfit, contrast0, contrast1)
}
\arguments{
\item{zlmfit}{ZlmFit output}

\item{contrast0}{vector of coefficients giving baseline contrast.  If missing, then the '(Intercept)' is used as baseline.}

\item{contrast1}{matrix of coefficients giving comparison contrasts.  If missing, then all non-(Intercept) coefficients are compared.}
}
\value{
list of matrices `logFC` and `varLogFC`, giving the log-fold-changes for each contrast (columns) and genes (rows) and the estimated sampling variance thereof
}
\description{
Using the delta method, estimate the log-fold change from a state given by a vector contrast0 and the state(s) given by contrast1
}
\section{Functions}{
\itemize{
\item \code{getLogFC}: Return results as a perhaps friendlier \code{data.table}
}}

