% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{lrTest,ZlmFit,character-method}
\alias{lrTest,ZlmFit,character-method}
\title{lrTest}
\usage{
\S4method{lrTest}{ZlmFit,character}(object, hypothesis)
}
\arguments{
\item{object}{ZlmFit}

\item{hypothesis}{See details}
}
\value{
3D array
}
\description{
A 3D array with first dimension being the genes,
next dimension giving information about the test
(the degrees of freedom, Chisq statistic, and P value), and final dimension
being the value of these quantities on the
discrete, continuous and hurdle (combined) levels.
}

