% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST-package.R
\name{MAST-defunct}
\alias{MAST-defunct}
\alias{filter}
\alias{cData}
\alias{fData}
\alias{exprs}
\title{Defunct functions in package `MAST`}
\description{
These functions are defunct or have been renamed.
}
\section{Functions (and replacements, if available)}{

\describe{
  \item{filter}{mast_filter}
  \item{\code{cData}}{\link{colData}}
  \item{\code{fData}}{\link{mcols}}
  \item{\code{exprs}}{\link{assay}}
  \item{\code{zlm.SingleCellAssay}}{\link{zlm}}
  \item{\code{combine}}{\link{cbind} or \link{rbind}}
  \item{\code{deviance_residuals_hook}}{No replacement available, underlying API changed}
}
}

