% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/LmWrapper.R
\docType{methods}
\name{model.matrix}
\alias{model.matrix}
\alias{model.matrix,LMlike-method}
\title{Model matrix accessor}
\usage{
model.matrix(object, ...)

\S4method{model.matrix}{LMlike}(object)
}
\arguments{
\item{object}{LMlike or subclass}
}
\value{
model.matrix if present
}
\description{
Model matrix accessor
}
\section{Methods (by class)}{
\itemize{
\item \code{LMlike}: return the \code{model.matrix}
}}

