% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fluidigm-methods.R
\name{numexp}
\alias{numexp}
\title{Report number of expressing cells per gene}
\usage{
numexp(sc)
}
\arguments{
\item{sc}{\code{SingleCellAssay}}
}
\value{
\code{numeric} vector
}
\description{
NAs are removed
}

