% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fluidigm-methods.R
\name{primerAverage}
\alias{primerAverage}
\title{Average expression values for duplicated/redundant genes}
\usage{
primerAverage(fd, geneGroups, fun.natural = expavg, fun.cycle = logshift)
}
\arguments{
\item{fd}{\code{SingleCellAssay} or subclass}

\item{geneGroups}{\code{character} naming a column in the \code{featureData} that keys the duplicates}

\item{fun.natural}{transformation to be used to collapse the duplicate expression values}

\item{fun.cycle}{transformation to be used after collapsing}
}
\value{
averaged version of \code{fd}.
}
\description{
Takes an average, potentially on a different scale given by \code{fun.natural}
of some genes. The average is then transformed with \code{fun.cycle}.
}
\note{
This code needs to be tested more extensively after a refactoring.  Caveat calculator.
}
