% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZlmFit.R
\name{print.summaryZlmFit}
\alias{print.summaryZlmFit}
\title{Print summary of a ZlmFit}
\usage{
\method{print}{summaryZlmFit}(x, n = 2, by = "logFC", ...)
}
\arguments{
\item{x}{output from summary(ZlmFit)}

\item{n}{number of genes to show}

\item{by}{one of 'C' , 'D' or 'logFC' for continuous, discrete and log fold change z-scores for each contrast}

\item{...}{ignored}
}
\value{
prints a pretty table and invisibly returns a \code{data.table} representing the table.
}
\description{
Shows the top `n` genes by z score on `by`
}
\seealso{
summary,ZlmFit-method
}
