% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SCASet-methods.R
\docType{methods}
\name{sampleNames,SCASet-method}
\alias{sampleNames,SCASet-method}
\title{get the sample names in an SCASet}
\usage{
\S4method{sampleNames}{SCASet}(object)
}
\arguments{
\item{object}{The \code{SCASet} object}
}
\value{
a \code{character} vector of the names of the samples in the \code{SCASet}
}
\description{
get the sample names in an SCASet
}

