% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{SCASet-class}
\alias{SCASet-class}
\alias{show,SCASet-method}
\title{SCASet is a set of SingleCellAssay objects or objects of its subclasses (i.e. FluidigmAssay)
The constructor \code{SCASet} should be used to make objects of this class.
}
\section{Slots}{
\describe{
\item{set}{A \code{list} of \code{SingleCellAssays} or its subclasses}
}}
\usage{
\S4method{show}{SCASet}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
SCASet is a set of SingleCellAssay objects or objects of its subclasses (i.e. FluidigmAssay)
The constructor \code{SCASet} should be used to make objects of this class.
}
\section{Slots}{
\describe{
\item{set}{A \code{list} of \code{SingleCellAssays} or its subclasses}
}
}

