% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zlmHooks.R
\name{score_hook}
\alias{score_hook}
\title{Used to compute module "scores"}
\usage{
score_hook(x)
}
\arguments{
\item{x}{the ZLMFit}
}
\description{
Hook to compute \eqn{ y_i-E(V_i)-\beta_{ngeneson} ng_i} for \eqn{y_i>0} and \eqn{E(U_i)(E(V_i)-\beta_{ngeneson} \times ng_i)} for \eqn{y_i=0}
}

