% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{se.coef}
\alias{se.coef}
\title{Return coefficient standard errors}
\usage{
se.coef(object, ...)
}
\arguments{
\item{object}{a model implementing \code{vcov}}

\item{...}{passed to methods}
}
\value{
vector or matrix
}
\description{
Given a fitted model, return the standard errors of the coefficient
}
\examples{
#see ZlmFit-class for examples
example('ZlmFit-class')
}
\seealso{
ZlmFit-class
}
