% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLayer.R
\docType{methods}
\name{show,DataLayer-method}
\alias{show,DataLayer-method}
\title{Show methods}
\usage{
\S4method{show}{DataLayer}(object)
}
\arguments{
\item{object}{whose human-readable form is desired.}
}
\description{
Show methods
}

