% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LmWrapper.R, R/ZlmFit.R
\name{show,LMlike-method}
\alias{show,LMlike-method}
\alias{show,ZlmFit-method}
\title{show}
\usage{
\S4method{show}{LMlike}(object)

\S4method{show}{ZlmFit}(object)
}
\arguments{
\item{object}{an object of some type}
}
\value{
side effect of printing to console
}
\description{
Display info
}
\details{
Prints information on a LMlike object
}
\section{Methods (by class)}{
\itemize{
\item \code{show(ZlmFit)}: print info on ZlmFit

}}
