% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{SingleCellAssay-package}
\alias{SingleCellAssay-package}
\title{"Tools for Single Cell Assay Analysis"}
\description{
Methods for analysing single cell assay data
}
\details{
This packages provides data structures and functions for statistical analysis of single-cell assay data such as Fluidigm single cell gene expression assays.
}
\seealso{
\code{\link{SCASet}}
}
\keyword{package}

