% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{split,DataLayer-method}
\alias{split,DataLayer-method}
\title{Split a DataLayer correctly}
\usage{
\S4method{split}{DataLayer}(x, f, drop = FALSE, ...)
}
\description{
split.default doesn't seem to do the right thing for an array
}

