% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[[,SCASet-method}
\alias{[[,SCASet,ANY-method}
\alias{[[,SCASet-method}
\title{Index into an SCASet object}
\usage{
\S4method{[[}{SCASet}(x, i, j, ...)
}
\arguments{
\item{x}{object to be subscripted}

\item{i}{index}

\item{j}{Ignored}

\item{...}{Ignored}
}
\value{
subscripted SingleCellAssay or derived class
}
\description{
Return a \code{SingleCellAssay} at index i
}

