% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLayer.R
\docType{methods}
\name{[[<-,DataLayer-method}
\alias{[[<-,DataLayer,ANY-method}
\alias{[[<-,DataLayer-method}
\title{Replace a section of a DataLayer}
\usage{
\S4method{[[}{DataLayer}(x, i, j, ...) <- value
}
\arguments{
\item{x}{DataLayer}

\item{i}{integer index(s)}

\item{j}{integer index(s)}

\item{...}{ignored}

\item{value}{a numeric to replace the selected indices}
}
\value{
DataLayer
}
\description{
Replace a section of a DataLayer
}

