% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{methods}
\name{subset}
\alias{subset}
\title{Evaluate an expression (returning a \code{logical} vector in the context of
\code{x}}
\usage{
subset(x, ...)
}
\arguments{
\item{x}{A matrix or \code{data.frame} typically}

\item{thesubset}{An expression, which evaluated in either the current
environment, or the parent, yields a logical vector}
}
\value{
The same class as \code{x} typically, with entries in which thesubset was \code{TRUE}
}
\description{
Really just syntactic sugar.
}
\keyword{transformation}

