% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAssay-methods.R
\docType{methods}
\name{subset,SingleCellAssay-method}
\alias{subset,SingleCellAssay-method}
\title{Subset a \code{SingleCellAssay} by cells (columns)}
\usage{
\S4method{subset}{SingleCellAssay}(x, ...)
}
\arguments{
\item{x}{\code{SingleCellAssay}}

\item{...}{expression}
}
\value{
\code{SingleCellAssay}
}
\description{
Evaluates the expression in \code{...} in the context of \code{colData(x)} and returns a subsetted version of \code{x}
}
\examples{
data(vbetaFA)
subset(vbetaFA, ncells==1)
}
