% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{subset}
\alias{subset}
\alias{subset,SingleCellAssay-method}
\title{Subset a SingleCellAssay by data in cellData}
\usage{
subset(x, ...)

\S4method{subset}{SingleCellAssay}(x, thesubset, ...)
}
\arguments{
\item{x}{SingleCellAssay}

\item{thesubset}{expression, which when evaluated in cellData environment which returns a logical}
}
\description{
Subset a SingleCellAssay by data in cellData
}
\examples{
data(vbetaFA)
subset(vbetaFA, ncells==1)
}

