% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{summarize}
\alias{summarize}
\title{Return programmatically useful summary of a fit}
\usage{
summarize(object, ...)
}
\arguments{
\item{object}{LMlike or subclass}

\item{...}{other arguments}
}
\value{
list of parameters characterizing fit
}
\description{
Return programmatically useful summary of a fit
}
