% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fluidigm-methods.R
\docType{methods}
\name{summary,SingleCellAssay-method}
\alias{summary,SingleCellAssay-method}
\title{Summarize expression parameters}
\usage{
\S4method{summary}{SingleCellAssay}(object, groups)
}
\arguments{
\item{groups}{character vector of grouping variables}

\item{fd}{object inheriting from SingleCellAssay}
}
\value{
list of mu, pi and num.
}
\description{
The mean of positive cells, mu, proportion of gene expression pi,
and number of expressing cells per \code{groups} per gene is returned as a list
}

