% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary}
\alias{summary}
\title{Return a human-readable summary}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{some object to be summarized}
}
\value{
a summary, possibly printed
}
\description{
Return a human-readable summary
}

