% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{thresholdNanoString}
\alias{thresholdNanoString}
\title{thresholds for positive expression}
\usage{
thresholdNanoString(nsa, ...)
}
\arguments{
\item{nsa}{NanostringAssay object}
}
\value{
modified nsa
}
\description{
thresholds for positive expression
}

