% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{thresholdNanoString}
\alias{thresholdNanoString}
\title{thresholds for positive expression}
\usage{
thresholdNanoString(nsa, ...)
}
\arguments{
\item{nsa}{NanostringAssay object}
}
\value{
modified nsa
}
\description{
thresholds for positive expression
}

