% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/thresholdSCRNA.R
\name{thresholdSCRNACountMatrix}
\alias{thresholdSCRNACountMatrix}
\title{Threshold a count matrix using an adaptive threshold.}
\usage{
thresholdSCRNACountMatrix(data_all, conditions = NULL, cutbins = NULL,
  nbins = 10, bin_by = "median", qt = 0.975, min_per_bin = 50,
  absolute_min = 0, return_log = TRUE, G = 2, plot = FALSE, adj = 1)
}
\arguments{
\item{data_all}{\code{matrix} of counts.  Rows are cells and columns are genes.}

\item{conditions}{Bins are be determined per gene and per condition.  Typically contrasts of interest should be specified.}

\item{cutbins}{\code{vector} of cut points.}

\item{nbins}{\code{integer} number of bins when cutbins is not specified.}

\item{bin_by}{\code{character} "median", "proportion", "mean"}

\item{qt}{when \code{bin_by} is "quantile", what quantile should be used to form the bins}

\item{min_per_bin}{minimum number of genes within a bin}

\item{absolute_min}{\code{numeric} giving a hard threshold below which everything is assumed to be noise}

\item{return_log}{return the logged expression matrix or not.  By default, returned expression matrix will be logged ( base 2 ).}

\item{G}{\code{integer} number of mixture model components to fit to the data above the threshold (default 2)}

\item{plot}{\code{logical} whether or not to plot thresholding results}

\item{adj}{\code{numeric} specifying the adjustment to the bandwidth}
}
\value{
\code{list} of thresholded counts (on natural scale), thresholds, bins, densities estimated on each bin, as well as observation weights from the mixture model fit, and the original data
}
\description{
An adaptive threshold is calculated from the conditional mean of expression, based on 10 bins
of the genes with similar expression levels. Thresholds are chosen by estimating cutpoints in the bimodal density estimates of the
binned data. You can adjust the bandwidth of the density estimation with \code{adj}. For each bin, the code will also fit a Gamma to the data on the left side of the cutpoint and a \code{G} component gaussian
mixture to the data on the right side of the cutpoint. The complete data in the bin is modeled as a mixture of these two, and observation weights from this mixture are returned as the \code{weights} component of the result.
If these are used for model fitting, the unthresholded data should be used. The code expects unlogged data as input.
}

