% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper-methods.R
\name{todt}
\alias{todt}
\title{Convert a pair of character names to a list that can be placed into a data.table}
\usage{
todt(v)
}
\arguments{
\item{v}{a character vector}
}
\value{
a call that can be evaluated
}
\description{
Convert a pair of character names to a list that can be placed into a data.table
}

