% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{vbetaFA}
\alias{vbetaFA}
\title{Vbeta Data Set, FluidigmAssay}
\format{a \code{FluidigmAssay} of the vbeta data set.}
\description{
Vbeta Data Set, FluidigmAssay
}
\seealso{
\code{\link{vbeta}}, \code{\link{FromFlatDF}}
}
