% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vcov}
\alias{vcov}
\title{Variance-covariance matrix for zero inflated}
\usage{
vcov(object, ...)
}
\arguments{
\item{object}{LMlike}

\item{which}{character vector, one of "C" (continuous) or "D" (discrete) specifying which component should be returned}
}
\value{
matrix
}
\description{
Given a fitted LMlike, return the variance-covariance from discrete or continuous
}

