% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{vcov}
\alias{vcov}
\title{Variance-covariance matrix for zero inflated}
\usage{
vcov(object, ...)
}
\arguments{
\item{object}{LMlike}

\item{...}{Additional parameters, namely \code{which}}
}
\value{
\code{matrix}
}
\description{
Given a fitted LMlike, return the variance-covariance from discrete or continuous.
\code{which}, a \code{character}, one of "C" (continuous) or "D" (discrete) must be specified, giving which component should be returned.
}

