% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultidimensionalScaling.R
\name{xform}
\alias{xform}
\title{Make matrix of continuous expression values, orthogonal to discrete}
\usage{
xform(mat, scale = FALSE)
}
\arguments{
\item{mat}{matrix (such as produced by exprs)}

\item{scale}{should the columns also be scaled to have unit variance}
}
\value{
\code{matrix}
}
\description{
This centers each column of \code{mat} around the mean of its non-zero values.
}
