% Generated by roxygen2 (4.0.1): do not edit by hand
\name{zlm.SingleCellAssay}
\alias{zlm.SingleCellAssay}
\title{zero-inflated regression for SingleCellAssay}
\usage{
zlm.SingleCellAssay(formula, sca, method = "glm", hypothesis, type = "Wald",
  keep.zlm = "false", .parallel = FALSE, silent = TRUE, ...)
}
\arguments{
\item{formula}{a formula with the measurement variable on the LHS and predictors present in cData on the RHS}

\item{sca}{SingleCellAssay object}

\item{method}{character vector, either 'glm' or 'glmer'}

\item{hypothesis}{character vector or list of character vectors passed to \code{lrTest} or \code{waldTest}.  See details.}

\item{type}{type of test to run, one of 'Wald' or 'LRT'}

\item{keep.zlm}{should the model objects be returned?  May be memory intensive.}

\item{silent}{Silence common problems with fitting some genes}

\item{fitArgs}{list of arguments passed to glm/glmer}
}
\value{
either an array of tests (one per primer), a list of such arrays (one per hypothesis),  or a list with components "models" and "fits".
}
\description{
For each gene in sca, fits the hurdle model in \code{formula} (linear for et>0), logistic for et==0 vs et>0.
Conducts tests specified in "hypothesis".
}
\details{
When keep.zlm is FALSE, a 3D array with first dimension being the genes,
next dimension giving information about the test
(the degrees of freedom, Chisq statistic, and P value), and final dimension
being the value of these quantities on the
discrete, continuous and hurdle (combined) levels.

When keep.zlm is TRUE, a list of length two is returned.
Component "tests" gives the above 3-D array.
Component "models" is a list giving the model fit for each gene.

When \code{hypothesis} is a list, then each test specified in the list will be run, and the returned object will also be a list of 3D arrays (or component "tests" will be a list if keep.zlm is TRUE).
}
\examples{
\dontrun{
data(vbetaFA)
testsByGene <- zlm.SingleCellAssay(~ Stim.Condition, vbetaFA, hypothesis='Stim.ConditionUnstim', method='glm', type='Wald')
# genes X metric X test type
dimnames(testsByGene)

modelsAndTestsByGene <- zlm.SingleCellAssay(~ Stim.Condition, vbeta.sc, hypothesis='Stim.ConditionUnstim', keep.zlm=TRUE)
names(modelsAndTestsByGene$models)
summary(modelsAndTestsByGene$models[['IL13']]$disc)
summary(modelsAndTestsByGene$models[['IL13']]$cont)

## Separate tests that Stim.Condition=0 and the Intercept=0
## (The second test doesn't make sense scientifically)
twoTests <- zlm.SingleCellAssay(~ Stim.Condition, vbeta.sc, hypothesis=list('Stim.ConditionUnstim', '(Intercept)'))
length(twoTests)
dimnames(twoTests[[1]])
}
}

