% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zlm.SingleCellAssay}
\alias{zlm.SingleCellAssay}
\title{Zero-inflated regression for SingleCellAssay}
\usage{
zlm.SingleCellAssay(formula, sca, method = "glm", silent = TRUE,
  ebayes = FALSE, ebayesControl = NULL, force = FALSE, hook = NULL,
  LMlike, ...)
}
\arguments{
\item{formula}{a formula with the measurement variable on the LHS and predictors present in cData on the RHS}

\item{sca}{SingleCellAssay object}

\item{method}{character vector, either 'glm', 'glmer' or 'bayesglm'}

\item{silent}{Silence common problems with fitting some genes}

\item{ebayes}{if TRUE, regularize variance using empirical bayes method}

\item{ebayesControl}{list with parameters for empirical bayes procedure.  See \link{ebayes}.}

\item{force}{Should we continue testing genes even after many errors have occurred?}

\item{hook}{a function called on the \code{fit} after each gene.}

\item{LMlike}{if provided, then the model defined in this object will be used, rather than following the formulas.  This is intended for internal use.}

\item{...}{arguments passed to the S4 model object.  For example, \code{fitArgsC} and \code{fitArgsD}.  These are a list of arguments passed to the underlying modeling functions.}
}
\value{
either an array of tests (one per primer), a list of such arrays (one per hypothesis),  or a list with components "models" and "fits".
}
\description{
For each gene in sca, fits the hurdle model in \code{formula} (linear for et>0), logistic for et==0 vs et>0.
Return an object of class \code{ZlmFit} containing slots giving the coefficients, variance-covariance matrices, etc.
After each gene, optionally run the function on the fit named by 'hook'
}
\details{
The empirical bayes regularization of the gene variance assumes that the precision (1/variance) is drawn from a
gamma distribution with unknown parameters.
These parameters are estimated by considering the distribution of sample variances over all genes.
The procedure used for this is determined from
\code{ebayesControl}, a named list with components 'method' (one of 'MOM' or 'MLE') and 'model' (one of 'H0' or 'H1')
method MOM uses a method-of-moments estimator, while MLE using the marginal likelihood.
H0 model estimates the precisions using the intercept alone in each gene, while H1 fits the full model specified by \code{formula}
}
\examples{
\dontrun{
data(vbetaFA)
zlmVbeta <- zlm.SingleCellAssay(~ Stim.Condition, vbeta.sc)
slotNames(zlmVbeta)
coef(zlmVbeta, 'C')['IL13',]

vcov(zlmVbeta, 'C')['IL13',,]
waldTest(zlmVbeta, CoefficientHypothesis('Stim.ConditionUnstim'))
}
}
\seealso{
ebayes, glmlike-class
}

